moved layout code to separate classes
