+++ src/engraving/libmscore/layout.cpp
@@ -97,6 +97,16 @@ namespace Ms {
 #endif
 
 //---------------------------------------------------------
+//   almostZero
+//---------------------------------------------------------
+
+static bool inline almostZero(qreal value)
+{
+    // 1e-3 is close enough to zero to see it as zero.
+    return value > -1e-3 && value < 1e-3;
+}
+
+//---------------------------------------------------------
 //   rebuildBspTree
 //---------------------------------------------------------
 
@@ -1274,79 +1284,6 @@ void Score::layoutChords3(std::vector<Note*>& notes, const Staff* staff, Segment
     }
 }
 
-#define beamModeMid(a) (a == Beam::Mode::MID || a == Beam::Mode::BEGIN32 || a == Beam::Mode::BEGIN64)
-
-bool beamNoContinue(Beam::Mode mode)
-{
-    return mode == Beam::Mode::END || mode == Beam::Mode::NONE || mode == Beam::Mode::INVALID;
-}
-
-//---------------------------------------------------------
-//   beamGraceNotes
-//---------------------------------------------------------
-
-void Score::beamGraceNotes(Chord* mainNote, bool after)
-{
-    ChordRest* a1    = 0;        // start of (potential) beam
-    Beam* beam       = 0;        // current beam
-    Beam::Mode bm = Beam::Mode::AUTO;
-    QVector<Chord*> graceNotes = after ? mainNote->graceNotesAfter() : mainNote->graceNotesBefore();
-
-    for (ChordRest* cr : qAsConst(graceNotes)) {
-        bm = Groups::endBeam(cr);
-        if ((cr->durationType().type() <= TDuration::DurationType::V_QUARTER) || (bm == Beam::Mode::NONE)) {
-            if (beam) {
-                beam->layoutGraceNotes();
-                beam = 0;
-            }
-            if (a1) {
-                a1->removeDeleteBeam(false);
-                a1 = 0;
-            }
-            cr->removeDeleteBeam(false);
-            continue;
-        }
-        if (beam) {
-            bool beamEnd = bm == Beam::Mode::BEGIN;
-            if (!beamEnd) {
-                cr->replaceBeam(beam);
-                cr = 0;
-                beamEnd = (bm == Beam::Mode::END);
-            }
-            if (beamEnd) {
-                beam->layoutGraceNotes();
-                beam = 0;
-            }
-        }
-        if (!cr) {
-            continue;
-        }
-        if (a1 == 0) {
-            a1 = cr;
-        } else {
-            if (!beamModeMid(bm) && (bm == Beam::Mode::BEGIN)) {
-                a1->removeDeleteBeam(false);
-                a1 = cr;
-            } else {
-                beam = a1->beam();
-                if (beam == 0 || beam->elements().front() != a1) {
-                    beam = new Beam(this);
-                    beam->setGenerated(true);
-                    beam->setTrack(mainNote->track());
-                    a1->replaceBeam(beam);
-                }
-                cr->replaceBeam(beam);
-                a1 = 0;
-            }
-        }
-    }
-    if (beam) {
-        beam->layoutGraceNotes();
-    } else if (a1) {
-        a1->removeDeleteBeam(false);
-    }
-}
-
 #if 0 // unused
 //---------------------------------------------------------
 //   layoutSpanner
@@ -1384,110 +1321,6 @@ void Score::layoutSpanner()
 }
 
 #endif
-
-//---------------------------------------------------------
-//   hideEmptyStaves
-//---------------------------------------------------------
-
-void Score::hideEmptyStaves(System* system, bool isFirstSystem)
-{
-    int staves   = _staves.size();
-    int staffIdx = 0;
-    bool systemIsEmpty = true;
-
-    for (Staff* staff : qAsConst(_staves)) {
-        SysStaff* ss  = system->staff(staffIdx);
-
-        Staff::HideMode hideMode = staff->hideWhenEmpty();
-
-        if (hideMode == Staff::HideMode::ALWAYS
-            || (styleB(Sid::hideEmptyStaves)
-                && (staves > 1)
-                && !(isFirstSystem && styleB(Sid::dontHideStavesInFirstSystem))
-                && hideMode != Staff::HideMode::NEVER)) {
-            bool hideStaff = true;
-            for (MeasureBase* m : system->measures()) {
-                if (!m->isMeasure()) {
-                    continue;
-                }
-                Measure* measure = toMeasure(m);
-                if (!measure->isEmpty(staffIdx)) {
-                    hideStaff = false;
-                    break;
-                }
-            }
-            // check if notes moved into this staff
-            Part* part = staff->part();
-            int n = part->nstaves();
-            if (hideStaff && (n > 1)) {
-                int idx = part->staves()->front()->idx();
-                for (int i = 0; i < part->nstaves(); ++i) {
-                    int st = idx + i;
-
-                    for (MeasureBase* mb : system->measures()) {
-                        if (!mb->isMeasure()) {
-                            continue;
-                        }
-                        Measure* m = toMeasure(mb);
-                        if (staff->hideWhenEmpty() == Staff::HideMode::INSTRUMENT && !m->isEmpty(st)) {
-                            hideStaff = false;
-                            break;
-                        }
-                        for (Segment* s = m->first(SegmentType::ChordRest); s; s = s->next(SegmentType::ChordRest)) {
-                            for (int voice = 0; voice < VOICES; ++voice) {
-                                ChordRest* cr = s->cr(st * VOICES + voice);
-                                if (cr == 0 || cr->isRest()) {
-                                    continue;
-                                }
-                                int staffMove = cr->staffMove();
-                                if (staffIdx == st + staffMove) {
-                                    hideStaff = false;
-                                    break;
-                                }
-                            }
-                        }
-                        if (!hideStaff) {
-                            break;
-                        }
-                    }
-                    if (!hideStaff) {
-                        break;
-                    }
-                }
-            }
-            ss->setShow(hideStaff ? false : staff->show());
-            if (ss->show()) {
-                systemIsEmpty = false;
-            }
-        } else if (!staff->show()) {
-            // TODO: OK to check this first and not bother with checking if empty?
-            ss->setShow(false);
-        } else {
-            systemIsEmpty = false;
-            ss->setShow(true);
-        }
-
-        ++staffIdx;
-    }
-    Staff* firstVisible = nullptr;
-    if (systemIsEmpty) {
-        for (Staff* staff : qAsConst(_staves)) {
-            SysStaff* ss  = system->staff(staff->idx());
-            if (staff->showIfEmpty() && !ss->show()) {
-                ss->setShow(true);
-                systemIsEmpty = false;
-            } else if (!firstVisible && staff->show()) {
-                firstVisible = staff;
-            }
-        }
-    }
-    // don’t allow a complete empty system
-    if (systemIsEmpty && !_staves.isEmpty()) {
-        Staff* staff = firstVisible ? firstVisible : _staves.front();
-        SysStaff* ss = system->staff(staff->idx());
-        ss->setShow(true);
-    }
-}
 
 //---------------------------------------------------------
 //   connectTies
@@ -1618,375 +1451,6 @@ void Score::relayoutForStyles()
 }
 
 //---------------------------------------------------------
-//   checkDivider
-//---------------------------------------------------------
-
-static void checkDivider(bool left, System* s, qreal yOffset, bool remove = false)
-{
-    SystemDivider* divider = left ? s->systemDividerLeft() : s->systemDividerRight();
-    if ((s->score()->styleB(left ? Sid::dividerLeft : Sid::dividerRight)) && !remove) {
-        if (!divider) {
-            divider = new SystemDivider(s->score());
-            divider->setDividerType(left ? SystemDivider::Type::LEFT : SystemDivider::Type::RIGHT);
-            divider->setGenerated(true);
-            s->add(divider);
-        }
-        divider->layout();
-        divider->rypos() = divider->height() * .5 + yOffset;
-        if (left) {
-            divider->rypos() += s->score()->styleD(Sid::dividerLeftY) * SPATIUM20;
-            divider->rxpos() =  s->score()->styleD(Sid::dividerLeftX) * SPATIUM20;
-        } else {
-            divider->rypos() += s->score()->styleD(Sid::dividerRightY) * SPATIUM20;
-            divider->rxpos() =  s->score()->styleD(Sid::pagePrintableWidth) * DPI - divider->width();
-            divider->rxpos() += s->score()->styleD(Sid::dividerRightX) * SPATIUM20;
-        }
-    } else if (divider) {
-        if (divider->generated()) {
-            s->remove(divider);
-            delete divider;
-        } else {
-            s->score()->undoRemoveElement(divider);
-        }
-    }
-}
-
-//---------------------------------------------------------
-//   almostZero
-//---------------------------------------------------------
-
-bool inline almostZero(qreal value)
-{
-    // 1e-3 is close enough to zero to see it as zero.
-    return value > -1e-3 && value < 1e-3;
-}
-
-//---------------------------------------------------------
-//   distributeStaves
-//---------------------------------------------------------
-
-static void distributeStaves(Page* page)
-{
-    Score* score { page->score() };
-    VerticalGapDataList vgdl;
-
-    // Find and classify all gaps between staves.
-    int ngaps { 0 };
-    qreal prevYBottom  { page->tm() };
-    qreal yBottom      { 0.0 };
-    qreal spacerOffset { 0.0 };
-    bool vbox          { false };
-    Spacer* nextSpacer { nullptr };
-    bool transferNormalBracket { false };
-    bool transferCurlyBracket  { false };
-    for (System* system : page->systems()) {
-        if (system->vbox()) {
-            VerticalGapData* vgd = new VerticalGapData(!ngaps++, system, nullptr, nullptr, nullptr, prevYBottom);
-            vgd->addSpaceAroundVBox(true);
-            prevYBottom = system->y();
-            yBottom     = system->y() + system->height();
-            vbox        = true;
-            vgdl.append(vgd);
-            transferNormalBracket = false;
-            transferCurlyBracket  = false;
-        } else {
-            bool newSystem       { true };
-            bool addSpaceAroundNormalBracket { false };
-            bool addSpaceAroundCurlyBracket  { false };
-            int endNormalBracket { -1 };
-            int endCurlyBracket  { -1 };
-            int staffNr { -1 };
-            for (SysStaff* sysStaff : *system->staves()) {
-                Staff* staff { score->staff(++staffNr) };
-                addSpaceAroundNormalBracket |= endNormalBracket == staffNr;
-                addSpaceAroundCurlyBracket  |= endCurlyBracket == staffNr;
-                for (const BracketItem* bi : staff->brackets()) {
-                    if (bi->bracketType() == BracketType::NORMAL) {
-                        addSpaceAroundNormalBracket |= staff->idx() > (endNormalBracket - 1);
-                        endNormalBracket = qMax(endNormalBracket, staff->idx() + bi->bracketSpan());
-                    } else if (bi->bracketType() == BracketType::BRACE) {
-                        addSpaceAroundCurlyBracket |= staff->idx() > (endCurlyBracket - 1);
-                        endCurlyBracket = qMax(endCurlyBracket, staff->idx() + bi->bracketSpan());
-                    }
-                }
-
-                if (!sysStaff->show()) {
-                    continue;
-                }
-
-                VerticalGapData* vgd = new VerticalGapData(!ngaps++, system, staff, sysStaff, nextSpacer, prevYBottom);
-                nextSpacer = system->downSpacer(staff->idx());
-
-                if (newSystem) {
-                    vgd->addSpaceBetweenSections();
-                    newSystem = false;
-                }
-                if (addSpaceAroundNormalBracket || transferNormalBracket) {
-                    vgd->addSpaceAroundNormalBracket();
-                    addSpaceAroundNormalBracket = false;
-                    transferNormalBracket = false;
-                }
-                if (addSpaceAroundCurlyBracket || transferCurlyBracket) {
-                    vgd->addSpaceAroundCurlyBracket();
-                    addSpaceAroundCurlyBracket = false;
-                    transferCurlyBracket = false;
-                } else if (staffNr < endCurlyBracket) {
-                    vgd->insideCurlyBracket();
-                }
-
-                if (vbox) {
-                    vgd->addSpaceAroundVBox(false);
-                    vbox = false;
-                }
-
-                prevYBottom  = system->y() + sysStaff->y() + sysStaff->bbox().height();
-                yBottom      = system->y() + sysStaff->y() + sysStaff->skyline().south().max();
-                spacerOffset = sysStaff->skyline().south().max() - sysStaff->bbox().height();
-                vgdl.append(vgd);
-            }
-            transferNormalBracket = endNormalBracket >= 0;
-            transferCurlyBracket  = endCurlyBracket >= 0;
-        }
-    }
-    --ngaps;
-
-    qreal spaceLeft { page->height() - page->bm() - score->styleP(Sid::staffLowerBorder) - yBottom };
-    if (nextSpacer) {
-        spaceLeft -= qMax(0.0, nextSpacer->gap() - spacerOffset - score->styleP(Sid::staffLowerBorder));
-    }
-    if (spaceLeft <= 0.0) {
-        return;
-    }
-
-    // Try to make the gaps equal, taking the spread factors and maximum spacing into account.
-    static const int maxPasses { 20 };     // Saveguard to prevent endless loops.
-    int pass { 0 };
-    while (!almostZero(spaceLeft) && (ngaps > 0) && (++pass < maxPasses)) {
-        ngaps = 0;
-        qreal smallest     { vgdl.smallest() };
-        qreal nextSmallest { vgdl.smallest(smallest) };
-        if (almostZero(smallest) || almostZero(nextSmallest)) {
-            break;
-        }
-
-        if ((nextSmallest - smallest) * vgdl.sumStretchFactor() > spaceLeft) {
-            nextSmallest = smallest + spaceLeft / vgdl.sumStretchFactor();
-        }
-
-        qreal addedSpace { 0.0 };
-        VerticalGapDataList modified;
-        for (VerticalGapData* vgd : vgdl) {
-            if (!almostZero(vgd->spacing() - smallest)) {
-                continue;
-            }
-            qreal step { nextSmallest - vgd->spacing() };
-            if (step < 0.0) {
-                continue;
-            }
-            step = vgd->addSpacing(step);
-            if (!almostZero(step)) {
-                addedSpace += step * vgd->factor();
-                modified.append(vgd);
-                ++ngaps;
-            }
-            if ((spaceLeft - addedSpace) <= 0.0) {
-                break;
-            }
-        }
-        if ((spaceLeft - addedSpace) <= 0.0) {
-            for (VerticalGapData* vgd : modified) {
-                vgd->undoLastAddSpacing();
-            }
-            ngaps = 0;
-        } else {
-            spaceLeft -= addedSpace;
-        }
-    }
-
-    // If there is still space left, distribute the space of the staves.
-    // However, there is a limit on how much space is added per gap.
-    const qreal maxPageFill { score->styleP(Sid::maxPageFillSpread) };
-    spaceLeft = qMin(maxPageFill * vgdl.length(), spaceLeft);
-    pass = 0;
-    ngaps = 1;
-    while (!almostZero(spaceLeft) && !almostZero(maxPageFill) && (ngaps > 0) && (++pass < maxPasses)) {
-        ngaps = 0;
-        qreal addedSpace { 0.0 };
-        qreal step { spaceLeft / vgdl.sumStretchFactor() };
-        for (VerticalGapData* vgd : vgdl) {
-            qreal res { vgd->addFillSpacing(step, maxPageFill) };
-            if (!almostZero(res)) {
-                addedSpace += res * vgd->factor();
-                ++ngaps;
-            }
-        }
-        spaceLeft -= addedSpace;
-    }
-
-    QSet<System*> systems;
-    qreal systemShift { 0.0 };
-    qreal staffShift  { 0.0 };
-    System* prvSystem { nullptr };
-    for (VerticalGapData* vgd : vgdl) {
-        if (vgd->sysStaff) {
-            systems.insert(vgd->system);
-        }
-        systemShift += vgd->actualAddedSpace();
-        if (prvSystem == vgd->system) {
-            staffShift += vgd->actualAddedSpace();
-        } else {
-            vgd->system->rypos() += systemShift;
-            if (prvSystem) {
-                prvSystem->setDistance(vgd->system->y() - prvSystem->y());
-                prvSystem->setHeight(prvSystem->height() + staffShift);
-            }
-            staffShift = 0.0;
-        }
-
-        if (vgd->sysStaff) {
-            vgd->sysStaff->bbox().translate(0.0, staffShift);
-        }
-
-        prvSystem = vgd->system;
-    }
-    if (prvSystem) {
-        prvSystem->setHeight(prvSystem->height() + staffShift);
-    }
-
-    for (System* system : systems) {
-        system->setMeasureHeight(system->height());
-        system->layoutBracketsVertical();
-        system->layoutInstrumentNames();
-    }
-    vgdl.deleteAll();
-}
-
-//---------------------------------------------------------
-//   layoutPage
-//    restHeight - vertical space which has to be distributed
-//                 between systems
-//    The algorithm tries to produce most equally spaced
-//    systems.
-//---------------------------------------------------------
-
-static void layoutPage(Page* page, qreal restHeight)
-{
-    if (restHeight < 0.0) {
-        qDebug("restHeight < 0.0: %f\n", restHeight);
-        restHeight = 0;
-    }
-
-    Score* score = page->score();
-    int gaps     = page->systems().size() - 1;
-
-    QList<System*> sList;
-
-    // build list of systems (excluding last)
-    // set initial distance for each to the unstretched minimum distance to next
-    for (int i = 0; i < gaps; ++i) {
-        System* s1 = page->systems().at(i);
-        System* s2 = page->systems().at(i + 1);
-        s1->setDistance(s2->y() - s1->y());
-        if (s1->vbox() || s2->vbox() || s1->hasFixedDownDistance()) {
-            if (s2->vbox()) {
-                checkDivider(true, s1, 0.0, true);              // remove
-                checkDivider(false, s1, 0.0, true);             // remove
-                checkDivider(true, s2, 0.0, true);              // remove
-                checkDivider(false, s2, 0.0, true);             // remove
-            }
-            continue;
-        }
-        sList.push_back(s1);
-    }
-
-    // last system needs no divider
-    System* lastSystem = page->systems().back();
-    checkDivider(true, lastSystem, 0.0, true);        // remove
-    checkDivider(false, lastSystem, 0.0, true);       // remove
-
-    if (sList.empty() || MScore::noVerticalStretch || score->enableVerticalSpread() || score->layoutMode() == LayoutMode::SYSTEM) {
-        if (score->layoutMode() == LayoutMode::FLOAT) {
-            qreal y = restHeight * .5;
-            for (System* system : page->systems()) {
-                system->move(PointF(0.0, y));
-            }
-        } else if ((score->layoutMode() != LayoutMode::SYSTEM) && score->enableVerticalSpread()) {
-            distributeStaves(page);
-        }
-
-        // system dividers
-        for (int i = 0; i < gaps; ++i) {
-            System* s1 = page->systems().at(i);
-            System* s2 = page->systems().at(i + 1);
-            if (!(s1->vbox() || s2->vbox())) {
-                qreal yOffset = s1->height() + (s1->distance() - s1->height()) * .5;
-                checkDivider(true,  s1, yOffset);
-                checkDivider(false, s1, yOffset);
-            }
-        }
-        return;
-    }
-
-    qreal maxDist = score->maxSystemDistance();
-
-    // allocate space as needed to normalize system distance (bottom of one system to top of next)
-    std::sort(sList.begin(), sList.end(), [](System* a, System* b) { return a->distance() - a->height() < b->distance() - b->height(); });
-    System* s0 = sList[0];
-    qreal dist = s0->distance() - s0->height();             // distance for shortest system
-    for (int i = 1; i < sList.size(); ++i) {
-        System* si = sList[i];
-        qreal ndist = si->distance() - si->height();        // next taller system
-        qreal fill  = ndist - dist;                         // amount by which this system distance exceeds next shorter
-        if (fill > 0.0) {
-            qreal totalFill = fill * i;                     // space required to add this amount to all shorter systems
-            if (totalFill > restHeight) {
-                totalFill = restHeight;                     // too much; adjust amount
-                fill = restHeight / i;
-            }
-            for (int k = 0; k < i; ++k) {                   // add amount to all shorter systems
-                System* s = sList[k];
-                qreal d = s->distance() + fill;
-                if ((d - s->height()) > maxDist) {          // but don't exceed max system distance
-                    d = qMax(maxDist + s->height(), s->distance());
-                }
-                s->setDistance(d);
-            }
-            restHeight -= totalFill;                        // reduce available space for next iteration
-            if (restHeight <= 0) {
-                break;                                      // no space left
-            }
-        }
-        dist = ndist;                                       // set up for next iteration
-    }
-
-    if (restHeight > 0.0) {                                 // space left?
-        qreal fill = restHeight / sList.size();
-        for (System* s : qAsConst(sList)) {                           // allocate it to systems equally
-            qreal d = s->distance() + fill;
-            if ((d - s->height()) > maxDist) {              // but don't exceed max system distance
-                d = qMax(maxDist + s->height(), s->distance());
-            }
-            s->setDistance(d);
-        }
-    }
-
-    qreal y = page->systems().at(0)->y();
-    for (int i = 0; i < gaps; ++i) {
-        System* s1  = page->systems().at(i);
-        System* s2  = page->systems().at(i + 1);
-        s1->rypos() = y;
-        y          += s1->distance();
-
-        if (!(s1->vbox() || s2->vbox())) {
-            qreal yOffset = s1->height() + (s1->distance() - s1->height()) * .5;
-            checkDivider(true,  s1, yOffset);
-            checkDivider(false, s1, yOffset);
-        }
-    }
-    page->systems().back()->rypos() = y;
-}
-
-//---------------------------------------------------------
 //   Spring
 //---------------------------------------------------------
 
@@ -2094,76 +1558,6 @@ void Score::respace(std::vector<ChordRest*>* elements)
         qreal dx = x - cr->segment()->pos().x();
         cr->rxpos() += dx;
     }
-}
-
-//---------------------------------------------------------
-//   getNextPage
-//---------------------------------------------------------
-
-void LayoutContext::getNextPage()
-{
-    if (!page || curPage >= score->npages()) {
-        page = new Page(score);
-        score->pages().push_back(page);
-        prevSystem = nullptr;
-        pageOldMeasure = nullptr;
-    } else {
-        page = score->pages()[curPage];
-        QList<System*>& systems = page->systems();
-        pageOldMeasure = systems.isEmpty() ? nullptr : systems.back()->measures().back();
-        const int i = systems.indexOf(curSystem);
-        if (i > 0 && systems[i - 1]->page() == page) {
-            // Current and previous systems are on the current page.
-            // Erase only the current and the following systems
-            // as the previous one will not participate in layout.
-            systems.erase(systems.begin() + i, systems.end());
-        } else { // system is not on the current page (or will be the first one)
-            systems.clear();
-        }
-        prevSystem = systems.empty() ? nullptr : systems.back();
-    }
-    page->bbox().setRect(0.0, 0.0, score->loWidth(), score->loHeight());
-    page->setNo(curPage);
-    qreal x = 0.0;
-    qreal y = 0.0;
-    if (curPage) {
-        Page* prevPage = score->pages()[curPage - 1];
-        if (MScore::verticalOrientation()) {
-            y = prevPage->pos().y() + page->height() + MScore::verticalPageGap;
-        } else {
-            qreal gap = (curPage + score->pageNumberOffset()) & 1 ? MScore::horizontalPageGapOdd : MScore::horizontalPageGapEven;
-            x = prevPage->pos().x() + page->width() + gap;
-        }
-    }
-    ++curPage;
-    page->setPos(x, y);
-}
-
-//---------------------------------------------------------
-//   getNextSystem
-//---------------------------------------------------------
-
-System* Score::getNextSystem(LayoutContext& lc)
-{
-    bool isVBox = lc.curMeasure->isVBox();
-    System* system;
-    if (lc.systemList.empty()) {
-        system = new System(this);
-        lc.systemOldMeasure = 0;
-    } else {
-        system = lc.systemList.takeFirst();
-        lc.systemOldMeasure = system->measures().empty() ? 0 : system->measures().back();
-        system->clear();       // remove measures from system
-    }
-    _systems.append(system);
-    if (!isVBox) {
-        int nstaves = Score::nstaves();
-        system->adjustStavesNumber(nstaves);
-        for (int i = 0; i < nstaves; ++i) {
-            system->staff(i)->setShow(score()->staff(i)->show());
-        }
-    }
-    return system;
 }
 
 //---------------------------------------------------------
@@ -2507,835 +1901,6 @@ void Score::createMMRest(Measure* firstMeasure, Measure* lastMeasure, const Frac
 }
 
 //---------------------------------------------------------
-// validMMRestMeasure
-//    return true if this might be a measure in a
-//    multi measure rest
-//---------------------------------------------------------
-
-static bool validMMRestMeasure(Measure* m)
-{
-    if (m->irregular()) {
-        return false;
-    }
-
-    int n = 0;
-    for (Segment* s = m->first(); s; s = s->next()) {
-        for (Element* e : s->annotations()) {
-            if (!(e->isRehearsalMark() || e->isTempoText() || e->isHarmony() || e->isStaffText() || e->isSystemText()
-                  || e->isInstrumentChange())) {
-                return false;
-            }
-        }
-        if (s->isChordRestType()) {
-            bool restFound = false;
-            int tracks = m->score()->ntracks();
-            for (int track = 0; track < tracks; ++track) {
-                if ((track % VOICES) == 0 && !m->score()->staff(track / VOICES)->show()) {
-                    track += VOICES - 1;
-                    continue;
-                }
-                if (s->element(track)) {
-                    if (!s->element(track)->isRest()) {
-                        return false;
-                    }
-                    restFound = true;
-                }
-            }
-            for (Element* e : s->annotations()) {
-                if (e->isFermata()) {
-                    return false;
-                }
-            }
-            if (restFound) {
-                ++n;
-            }
-            // measure is not empty if there is more than one rest
-            if (n > 1) {
-                return false;
-            }
-        }
-    }
-    return true;
-}
-
-//---------------------------------------------------------
-//  breakMultiMeasureRest
-//    return true if this measure should start a new
-//    multi measure rest
-//---------------------------------------------------------
-
-static bool breakMultiMeasureRest(Measure* m)
-{
-    if (m->breakMultiMeasureRest()) {
-        return true;
-    }
-
-    if (m->repeatStart()
-        || (m->prevMeasure() && m->prevMeasure()->repeatEnd())
-        || (m->isIrregular())
-        || (m->prevMeasure() && m->prevMeasure()->isIrregular())
-        || (m->prevMeasure() && (m->prevMeasure()->sectionBreak()))) {
-        return true;
-    }
-
-    auto sl = m->score()->spannerMap().findOverlapping(m->tick().ticks(), m->endTick().ticks());
-    for (auto i : sl) {
-        Spanner* s = i.value;
-        // break for first measure of volta or textline and first measure *after* volta
-        if ((s->isVolta() || s->isTextLine()) && (s->tick() == m->tick() || s->tick2() == m->tick())) {
-            return true;
-        }
-    }
-
-    // break for marker in this measure
-    for (Element* e : m->el()) {
-        if (e->isMarker()) {
-            Marker* mark = toMarker(e);
-            if (!(mark->align() & Align::RIGHT)) {
-                return true;
-            }
-        }
-    }
-
-    // break for marker & jump in previous measure
-    Measure* pm = m->prevMeasure();
-    if (pm) {
-        for (Element* e : pm->el()) {
-            if (e->isJump()) {
-                return true;
-            } else if (e->isMarker()) {
-                Marker* mark = toMarker(e);
-                if (mark->align() & Align::RIGHT) {
-                    return true;
-                }
-            }
-        }
-    }
-
-    // break for MeasureRepeat group
-    for (int staffIdx = 0; staffIdx < m->score()->nstaves(); ++staffIdx) {
-        if (m->isMeasureRepeatGroup(staffIdx)
-            || (m->prevMeasure() && m->prevMeasure()->isMeasureRepeatGroup(staffIdx))) {
-            return true;
-        }
-    }
-
-    for (Segment* s = m->first(); s; s = s->next()) {
-        for (Element* e : s->annotations()) {
-            if (!e->visible()) {
-                continue;
-            }
-            if (e->isRehearsalMark()
-                || e->isTempoText()
-                || ((e->isHarmony() || e->isStaffText() || e->isSystemText() || e->isInstrumentChange())
-                    && (e->systemFlag() || m->score()->staff(e->staffIdx())->show()))) {
-                return true;
-            }
-        }
-        for (int staffIdx = 0; staffIdx < m->score()->nstaves(); ++staffIdx) {
-            if (!m->score()->staff(staffIdx)->show()) {
-                continue;
-            }
-            Element* e = s->element(staffIdx * VOICES);
-            if (!e || e->generated()) {
-                continue;
-            }
-            if (s->isStartRepeatBarLineType()) {
-                return true;
-            }
-            if (s->isType(SegmentType::KeySig | SegmentType::TimeSig) && m->tick().isNotZero()) {
-                return true;
-            }
-            if (s->isClefType()) {
-                if (s->tick() != m->endTick() && m->tick().isNotZero()) {
-                    return true;
-                }
-            }
-        }
-    }
-    if (pm) {
-        Segment* s = pm->findSegmentR(SegmentType::EndBarLine, pm->ticks());
-        if (s) {
-            for (int staffIdx = 0; staffIdx < s->score()->nstaves(); ++staffIdx) {
-                BarLine* bl = toBarLine(s->element(staffIdx * VOICES));
-                if (bl) {
-                    BarLineType t = bl->barLineType();
-                    if (t != BarLineType::NORMAL && t != BarLineType::BROKEN && t != BarLineType::DOTTED && !bl->generated()) {
-                        return true;
-                    } else {
-                        break;
-                    }
-                }
-            }
-        }
-        if (pm->findSegment(SegmentType::Clef, m->tick())) {
-            return true;
-        }
-    }
-    return false;
-}
-
-//---------------------------------------------------------
-//   adjustMeasureNo
-//---------------------------------------------------------
-
-int LayoutContext::adjustMeasureNo(MeasureBase* m)
-{
-    measureNo += m->noOffset();
-    m->setNo(measureNo);
-    if (!m->irregular()) {          // don’t count measure
-        ++measureNo;
-    }
-    if (m->sectionBreakElement() && m->sectionBreakElement()->startWithMeasureOne()) {
-        measureNo = 0;
-    }
-    return measureNo;
-}
-
-//---------------------------------------------------------
-//   createBeams
-//    helper function
-//---------------------------------------------------------
-
-void Score::createBeams(LayoutContext& lc, Measure* measure)
-{
-    bool crossMeasure = styleB(Sid::crossMeasureValues);
-
-    for (int track = 0; track < ntracks(); ++track) {
-        Staff* stf = staff(track2staff(track));
-
-        // don’t compute beams for invisible staves and tablature without stems
-        if (!stf->show() || (stf->isTabStaff(measure->tick()) && stf->staffType(measure->tick())->stemless())) {
-            continue;
-        }
-
-        ChordRest* a1    = 0;          // start of (potential) beam
-        bool firstCR     = true;
-        Beam* beam       = 0;          // current beam
-        Beam::Mode bm    = Beam::Mode::AUTO;
-        ChordRest* prev  = 0;
-        bool checkBeats  = false;
-        Fraction stretch = Fraction(1, 1);
-        QHash<int, TDuration> beatSubdivision;
-
-        // if this measure is simple meter (actually X/4),
-        // then perform a prepass to determine the subdivision of each beat
-
-        beatSubdivision.clear();
-        TimeSig* ts = stf->timeSig(measure->tick());
-        checkBeats  = false;
-        stretch     = ts ? ts->stretch() : Fraction(1, 1);
-
-        const SegmentType st = SegmentType::ChordRest;
-        if (ts && ts->denominator() == 4) {
-            checkBeats = true;
-            for (Segment* s = measure->first(st); s; s = s->next(st)) {
-                ChordRest* mcr = toChordRest(s->element(track));
-                if (mcr == 0) {
-                    continue;
-                }
-                int beat = (mcr->rtick() * stretch).ticks() / MScore::division;
-                if (beatSubdivision.contains(beat)) {
-                    beatSubdivision[beat] = qMin(beatSubdivision[beat], mcr->durationType());
-                } else {
-                    beatSubdivision[beat] = mcr->durationType();
-                }
-            }
-        }
-
-        for (Segment* segment = measure->first(st); segment; segment = segment->next(st)) {
-            ChordRest* cr = segment->cr(track);
-            if (cr == 0) {
-                continue;
-            }
-
-            if (firstCR) {
-                firstCR = false;
-                // Handle cross-measure beams
-                Beam::Mode mode = cr->beamMode();
-                if (mode == Beam::Mode::MID || mode == Beam::Mode::END) {
-                    ChordRest* prevCR = findCR(measure->tick() - Fraction::fromTicks(1), track);
-                    if (prevCR) {
-                        const Measure* pm = prevCR->measure();
-                        if (!beamNoContinue(prevCR->beamMode())
-                            && !pm->lineBreak() && !pm->pageBreak() && !pm->sectionBreak()
-                            && lc.prevMeasure
-                            && !(prevCR->isChord() && prevCR->durationType().type() <= TDuration::DurationType::V_QUARTER)) {
-                            beam = prevCR->beam();
-                            //a1 = beam ? beam->elements().front() : prevCR;
-                            a1 = beam ? nullptr : prevCR;               // when beam is found, a1 is no longer required.
-                        }
-                    }
-                }
-            }
-#if 0
-            for (Lyrics* l : cr->lyrics()) {
-                if (l) {
-                    l->layout();
-                }
-            }
-#endif
-            // handle grace notes and cross-measure beaming
-            // (tied chords?)
-            if (cr->isChord()) {
-                Chord* chord = toChord(cr);
-                beamGraceNotes(chord, false);         // grace before
-                beamGraceNotes(chord, true);          // grace after
-                // set up for cross-measure values as soon as possible
-                // to have all computations (stems, hooks, ...) consistent with it
-                if (!chord->isGrace()) {
-                    chord->crossMeasureSetup(crossMeasure);
-                }
-            }
-
-            if (cr->isRest() && cr->beamMode() == Beam::Mode::AUTO) {
-                bm = Beam::Mode::NONE;                   // do not beam rests set to Beam::Mode::AUTO
-            } else {
-                bm = Groups::endBeam(cr, prev);          // get defaults from time signature properties
-            }
-            // perform additional context-dependent checks
-            if (bm == Beam::Mode::AUTO) {
-                // check if we need to break beams according to minimum duration in current / previous beat
-                if (checkBeats && cr->rtick().isNotZero()) {
-                    Fraction tick = cr->rtick() * stretch;
-                    // check if on the beat
-                    if ((tick.ticks() % MScore::division) == 0) {
-                        int beat = tick.ticks() / MScore::division;
-                        // get minimum duration for this & previous beat
-                        TDuration minDuration = qMin(beatSubdivision[beat], beatSubdivision[beat - 1]);
-                        // re-calculate beam as if this were the duration of current chordrest
-                        TDuration saveDuration        = cr->actualDurationType();
-                        TDuration saveCMDuration      = cr->crossMeasureDurationType();
-                        CrossMeasure saveCrossMeasVal = cr->crossMeasure();
-                        cr->setDurationType(minDuration);
-                        bm = Groups::endBeam(cr, prev);
-                        cr->setDurationType(saveDuration);
-                        cr->setCrossMeasure(saveCrossMeasVal);
-                        cr->setCrossMeasureDurationType(saveCMDuration);
-                    }
-                }
-            }
-
-            prev = cr;
-
-            // if chord has hooks and is 2nd element of a cross-measure value
-            // set beam mode to NONE (do not combine with following chord beam/hook, if any)
-
-            if (cr->durationType().hooks() > 0 && cr->crossMeasure() == CrossMeasure::SECOND) {
-                bm = Beam::Mode::NONE;
-            }
-
-            if ((cr->isChord() && cr->durationType().type() <= TDuration::DurationType::V_QUARTER) || (bm == Beam::Mode::NONE)) {
-                bool removeBeam = true;
-                if (beam) {
-                    beam->layout1();
-                    removeBeam = (beam->elements().size() <= 1);
-                    beam = 0;
-                }
-                if (a1) {
-                    if (removeBeam) {
-                        a1->removeDeleteBeam(false);
-                    }
-                    a1 = 0;
-                }
-                cr->removeDeleteBeam(false);
-                continue;
-            }
-
-            if (beam) {
-                bool beamEnd = (bm == Beam::Mode::BEGIN);
-                if (!beamEnd) {
-                    cr->replaceBeam(beam);
-                    cr = 0;
-                    beamEnd = (bm == Beam::Mode::END);
-                }
-                if (beamEnd) {
-                    beam->layout1();
-                    beam = 0;
-                }
-            }
-            if (!cr) {
-                continue;
-            }
-
-            if (a1 == 0) {
-                a1 = cr;
-            } else {
-                if (!beamModeMid(bm)
-                    &&
-                    (bm == Beam::Mode::BEGIN
-                     || (a1->segment()->segmentType() != cr->segment()->segmentType())
-                     || (a1->tick() + a1->actualTicks() < cr->tick())
-                    )
-                    ) {
-                    a1->removeDeleteBeam(false);
-                    a1 = cr;
-                } else {
-                    beam = a1->beam();
-                    if (beam == 0 || beam->elements().front() != a1) {
-                        beam = new Beam(this);
-                        beam->setGenerated(true);
-                        beam->setTrack(track);
-                        a1->replaceBeam(beam);
-                    }
-                    cr->replaceBeam(beam);
-                    a1 = 0;
-                }
-            }
-        }
-        if (beam) {
-            beam->layout1();
-        } else if (a1) {
-            Fraction nextTick = a1->tick() + a1->actualTicks();
-            Measure* m = (nextTick >= measure->endTick() ? measure->nextMeasure() : measure);
-            ChordRest* nextCR = (m ? m->findChordRest(nextTick, track) : nullptr);
-            Beam* b = a1->beam();
-            if (!(b && b->elements().startsWith(a1) && nextCR && beamModeMid(nextCR->beamMode()))) {
-                a1->removeDeleteBeam(false);
-            }
-        }
-    }
-}
-
-//---------------------------------------------------------
-//   breakCrossMeasureBeams
-//---------------------------------------------------------
-
-static void breakCrossMeasureBeams(Measure* measure)
-{
-    MeasureBase* mbNext = measure->next();
-    if (!mbNext || !mbNext->isMeasure()) {
-        return;
-    }
-
-    Measure* next = toMeasure(mbNext);
-    Score* score = measure->score();
-    const int ntracks = score->ntracks();
-    Segment* fstSeg = next->first(SegmentType::ChordRest);
-    if (!fstSeg) {
-        return;
-    }
-
-    for (int track = 0; track < ntracks; ++track) {
-        Staff* stf = score->staff(track2staff(track));
-
-        // don’t compute beams for invisible staves and tablature without stems
-        if (!stf->show() || (stf->isTabStaff(measure->tick()) && stf->staffType(measure->tick())->stemless())) {
-            continue;
-        }
-
-        Element* e = fstSeg->element(track);
-        if (!e || !e->isChordRest()) {
-            continue;
-        }
-
-        ChordRest* cr = toChordRest(e);
-        Beam* beam = cr->beam();
-        if (!beam || beam->elements().front()->measure() == next) {   // no beam or not cross-measure beam
-            continue;
-        }
-
-        std::vector<ChordRest*> mElements;
-        std::vector<ChordRest*> nextElements;
-
-        for (ChordRest* beamCR : beam->elements()) {
-            if (beamCR->measure() == measure) {
-                mElements.push_back(beamCR);
-            } else {
-                nextElements.push_back(beamCR);
-            }
-        }
-
-        if (mElements.size() == 1) {
-            mElements[0]->removeDeleteBeam(false);
-        }
-
-        Beam* newBeam = nullptr;
-        if (nextElements.size() > 1) {
-            newBeam = new Beam(score);
-            newBeam->setGenerated(true);
-            newBeam->setTrack(track);
-        }
-
-        const bool nextBeamed = bool(newBeam);
-        for (ChordRest* nextCR : nextElements) {
-            nextCR->removeDeleteBeam(nextBeamed);
-            if (newBeam) {
-                newBeam->add(nextCR);
-            }
-        }
-
-        if (newBeam) {
-            newBeam->layout1();
-        }
-    }
-}
-
-//---------------------------------------------------------
-//   layoutDrumsetChord
-//---------------------------------------------------------
-
-void layoutDrumsetChord(Chord* c, const Drumset* drumset, const StaffType* st, qreal spatium)
-{
-    for (Note* note : c->notes()) {
-        int pitch = note->pitch();
-        if (!drumset->isValid(pitch)) {
-            // qDebug("unmapped drum note %d", pitch);
-        } else if (!note->fixed()) {
-            note->undoChangeProperty(Pid::HEAD_GROUP, int(drumset->noteHead(pitch)));
-            int line = drumset->line(pitch);
-            note->setLine(line);
-
-            int off  = st->stepOffset();
-            qreal ld = st->lineDistance().val();
-            note->rypos()  = (line + off * 2.0) * spatium * .5 * ld;
-        }
-    }
-}
-
-//---------------------------------------------------------
-//   extendedStemLenWithTwoNotesTremolo
-//    Goal: To extend stem of one of the chords to make the tremolo less steep
-//    Returns a modified pair of stem lengths of two chords
-//---------------------------------------------------------
-
-std::pair<qreal, qreal> extendedStemLenWithTwoNoteTremolo(Tremolo* tremolo, qreal stemLen1, qreal stemLen2)
-{
-    const qreal spatium = tremolo->spatium();
-    Chord* c1 = tremolo->chord1();
-    Chord* c2 = tremolo->chord2();
-    Stem* s1 = c1->stem();
-    Stem* s2 = c2->stem();
-    const qreal sgn1 = c1->up() ? -1.0 : 1.0;
-    const qreal sgn2 = c2->up() ? -1.0 : 1.0;
-    const qreal stemTipDistance = (s1 && s2) ? (s2->pagePos().y() + stemLen2) - (s1->pagePos().y() + stemLen1)
-                                  : (c2->stemPos().y() + stemLen2) - (c1->stemPos().y() + stemLen1);
-
-    // same staff & same direction: extend one of the stems
-    if (c1->staffMove() == c2->staffMove() && c1->up() == c2->up()) {
-        const bool stem1Higher = stemTipDistance > 0.0;
-        if (std::abs(stemTipDistance) > 1.0 * spatium) {
-            if ((c1->up() && !stem1Higher) || (!c1->up() && stem1Higher)) {
-                return { stemLen1 + sgn1 * (std::abs(stemTipDistance) - 1.0 * spatium), stemLen2 };
-            } else {   /* if ((c1->up() && stem1Higher) || (!c1->up() && !stem1Higher)) */
-                return { stemLen1, stemLen2 + sgn2 * (std::abs(stemTipDistance) - 1.0 * spatium) };
-            }
-        }
-    }
-
-// TODO: cross-staff two-note tremolo. Currently doesn't generate the right result in some cases.
-#if 0
-    // cross-staff & beam between staves: extend both stems by the same length
-    else if (tremolo->crossStaffBeamBetween()) {
-        const qreal sw = tremolo->score()->styleS(Sid::tremoloStrokeWidth).val();
-        const qreal td = tremolo->score()->styleS(Sid::tremoloDistance).val();
-        const qreal tremoloMinHeight = ((tremolo->lines() - 1) * td + sw) * spatium;
-        const qreal dy = c1->up() ? tremoloMinHeight - stemTipDistance : tremoloMinHeight + stemTipDistance;
-        const bool tooShort = dy > 1.0 * spatium;
-        const bool tooLong = dy < -1.0 * spatium;
-        const qreal idealDistance = 1.0 * spatium - tremoloMinHeight;
-
-        if (tooShort) {
-            return { stemLen1 + sgn1 * (std::abs(stemTipDistance) - idealDistance) / 2.0,
-                     stemLen2 + sgn2 * (std::abs(stemTipDistance) - idealDistance) / 2.0 };
-        } else if (tooLong) {
-            return { stemLen1 - sgn1 * (std::abs(stemTipDistance) + idealDistance) / 2.0,
-                     stemLen2 - sgn2 * (std::abs(stemTipDistance) + idealDistance) / 2.0 };
-        }
-    }
-#endif
-
-    return { stemLen1, stemLen2 };
-}
-
-//---------------------------------------------------------
-//   getNextMeasure
-//---------------------------------------------------------
-
-void Score::getNextMeasure(LayoutContext& lc)
-{
-    lc.prevMeasure = lc.curMeasure;
-    lc.curMeasure  = lc.nextMeasure;
-    if (!lc.curMeasure) {
-        lc.nextMeasure = _showVBox ? first() : firstMeasure();
-    } else {
-        lc.nextMeasure = _showVBox ? lc.curMeasure->next() : lc.curMeasure->nextMeasure();
-    }
-    if (!lc.curMeasure) {
-        return;
-    }
-
-    int mno = lc.adjustMeasureNo(lc.curMeasure);
-
-    if (lc.curMeasure->isMeasure()) {
-        if (score()->styleB(Sid::createMultiMeasureRests)) {
-            Measure* m = toMeasure(lc.curMeasure);
-            Measure* nm = m;
-            Measure* lm = nm;
-            int n       = 0;
-            Fraction len;
-
-            while (validMMRestMeasure(nm)) {
-                MeasureBase* mb = _showVBox ? nm->next() : nm->nextMeasure();
-                if (breakMultiMeasureRest(nm) && n) {
-                    break;
-                }
-                if (nm != m) {
-                    lc.adjustMeasureNo(nm);
-                }
-                ++n;
-                len += nm->ticks();
-                lm = nm;
-                if (!(mb && mb->isMeasure())) {
-                    break;
-                }
-                nm = toMeasure(mb);
-            }
-            if (n >= styleI(Sid::minEmptyMeasures)) {
-                createMMRest(m, lm, len);
-                lc.curMeasure  = m->mmRest();
-                lc.nextMeasure = _showVBox ? lm->next() : lm->nextMeasure();
-            } else {
-                if (m->mmRest()) {
-                    undo(new ChangeMMRest(m, 0));
-                }
-                m->setMMRestCount(0);
-                lc.measureNo = mno;
-            }
-        } else if (toMeasure(lc.curMeasure)->isMMRest()) {
-            qDebug("mmrest: no %d += %d", lc.measureNo, toMeasure(lc.curMeasure)->mmRestCount());
-            lc.measureNo += toMeasure(lc.curMeasure)->mmRestCount() - 1;
-        }
-    }
-    if (!lc.curMeasure->isMeasure()) {
-        lc.curMeasure->setTick(lc.tick);
-        return;
-    }
-
-    //-----------------------------------------
-    //    process one measure
-    //-----------------------------------------
-
-    Measure* measure = toMeasure(lc.curMeasure);
-    measure->moveTicks(lc.tick - measure->tick());
-
-    if (lineMode() && (measure->tick() < lc.startTick || measure->tick() > lc.endTick)) {
-        // needed to reset segment widths if they can change after measure width is computed
-        //for (Segment& s : measure->segments())
-        //      s.createShapes();
-        lc.tick += measure->ticks();
-        return;
-    }
-
-    measure->connectTremolo();
-
-    //
-    // calculate accidentals and note lines,
-    // create stem and set stem direction
-    //
-    for (int staffIdx = 0; staffIdx < score()->nstaves(); ++staffIdx) {
-        const Staff* staff     = Score::staff(staffIdx);
-        const Drumset* drumset
-            = staff->part()->instrument(measure->tick())->useDrumset() ? staff->part()->instrument(measure->tick())->drumset() : 0;
-        AccidentalState as;          // list of already set accidentals for this measure
-        as.init(staff->keySigEvent(measure->tick()), staff->clef(measure->tick()));
-
-        for (Segment& segment : measure->segments()) {
-            // TODO? maybe we do need to process it here to make it possible to enable later
-            //if (!segment.enabled())
-            //      continue;
-            if (segment.isKeySigType()) {
-                KeySig* ks = toKeySig(segment.element(staffIdx * VOICES));
-                if (!ks) {
-                    continue;
-                }
-                Fraction tick = segment.tick();
-                as.init(staff->keySigEvent(tick), staff->clef(tick));
-                ks->layout();
-            } else if (segment.isChordRestType()) {
-                const StaffType* st = staff->staffTypeForElement(&segment);
-                int track     = staffIdx * VOICES;
-                int endTrack  = track + VOICES;
-
-                for (int t = track; t < endTrack; ++t) {
-                    ChordRest* cr = segment.cr(t);
-                    if (!cr) {
-                        continue;
-                    }
-                    qreal m = staff->staffMag(&segment);
-                    if (cr->isSmall()) {
-                        m *= score()->styleD(Sid::smallNoteMag);
-                    }
-
-                    if (cr->isChord()) {
-                        Chord* chord = toChord(cr);
-                        chord->cmdUpdateNotes(&as);
-                        for (Chord* c : chord->graceNotes()) {
-                            c->setMag(m * score()->styleD(Sid::graceNoteMag));
-                            c->computeUp();
-                            if (c->stemDirection() != Direction::AUTO) {
-                                c->setUp(c->stemDirection() == Direction::UP);
-                            } else {
-                                c->setUp(!(t % 2));
-                            }
-                            if (drumset) {
-                                layoutDrumsetChord(c, drumset, st, spatium());
-                            }
-                            c->layoutStem1();
-                        }
-                        if (drumset) {
-                            layoutDrumsetChord(chord, drumset, st, spatium());
-                        }
-                        chord->computeUp();
-                        chord->layoutStem1();               // create stems needed to calculate spacing
-                                                            // stem direction can change later during beam processing
-
-                        // if there is a two-note tremolo attached, and it is too steep,
-                        // extend stem of one of the chords (if not cross-staff)
-                        // or extend both stems (if cross-staff)
-                        // this should be done after the stem lengths of two notes are both calculated
-                        if (chord->tremolo() && chord == chord->tremolo()->chord2()) {
-                            Stem* stem1 = chord->tremolo()->chord1()->stem();
-                            Stem* stem2 = chord->tremolo()->chord2()->stem();
-                            if (stem1 && stem2) {
-                                std::pair<qreal, qreal> extendedLen = extendedStemLenWithTwoNoteTremolo(
-                                    chord->tremolo(),
-                                    stem1->p2().y(),
-                                    stem2->p2().y());
-                                stem1->setLen(extendedLen.first);
-                                stem2->setLen(extendedLen.second);
-                            }
-                        }
-                    }
-                    cr->setMag(m);
-                }
-            } else if (segment.isClefType()) {
-                Element* e = segment.element(staffIdx * VOICES);
-                if (e) {
-                    toClef(e)->setSmall(true);
-                    e->layout();
-                }
-            } else if (segment.isType(SegmentType::TimeSig | SegmentType::Ambitus | SegmentType::HeaderClef)) {
-                Element* e = segment.element(staffIdx * VOICES);
-                if (e) {
-                    e->layout();
-                }
-            }
-        }
-    }
-
-    createBeams(lc, measure);
-
-    for (int staffIdx = 0; staffIdx < score()->nstaves(); ++staffIdx) {
-        for (Segment& segment : measure->segments()) {
-            if (segment.isChordRestType()) {
-                layoutChords1(&segment, staffIdx);
-                for (int voice = 0; voice < VOICES; ++voice) {
-                    ChordRest* cr = segment.cr(staffIdx * VOICES + voice);
-                    if (cr) {
-                        for (Lyrics* l : cr->lyrics()) {
-                            if (l) {
-                                l->layout();
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    measure->computeTicks();
-
-    for (Segment& segment : measure->segments()) {
-        if (segment.isBreathType()) {
-            for (Element* e : segment.elist()) {
-                if (e && e->isBreath()) {
-                    e->layout();
-                }
-            }
-        } else if (segment.isChordRestType()) {
-            for (Element* e : segment.annotations()) {
-                if (e->isSymbol()) {
-                    e->layout();
-                }
-            }
-        }
-    }
-
-    rebuildTempoAndTimeSigMaps(measure);
-
-    Segment* seg = measure->findSegmentR(SegmentType::StartRepeatBarLine, Fraction(0, 1));
-    if (measure->repeatStart()) {
-        if (!seg) {
-            seg = measure->getSegmentR(SegmentType::StartRepeatBarLine, Fraction(0, 1));
-        }
-        measure->barLinesSetSpan(seg);          // this also creates necessary barlines
-        for (int staffIdx = 0; staffIdx < nstaves(); ++staffIdx) {
-            BarLine* b = toBarLine(seg->element(staffIdx * VOICES));
-            if (b) {
-                b->setBarLineType(BarLineType::START_REPEAT);
-                b->layout();
-            }
-        }
-    } else if (seg) {
-        score()->undoRemoveElement(seg);
-    }
-
-    for (Segment& s : measure->segments()) {
-        // TODO? maybe we do need to process it here to make it possible to enable later
-        //if (!s.enabled())
-        //      continue;
-        // DEBUG: relayout grace notes as beaming/flags may have changed
-        if (s.isChordRestType()) {
-            for (Element* e : s.elist()) {
-                if (e && e->isChord()) {
-                    Chord* chord = toChord(e);
-                    chord->layout();
-//                              if (chord->tremolo())            // debug
-//                                    chord->tremolo()->layout();
-                }
-            }
-        } else if (s.isEndBarLineType()) {
-            continue;
-        }
-        s.createShapes();
-    }
-
-    lc.tick += measure->ticks();
-}
-
-//---------------------------------------------------------
-//   isTopBeam
-//    returns true for the first CR of a beam that is not cross-staff
-//---------------------------------------------------------
-
-bool isTopBeam(ChordRest* cr)
-{
-    Beam* b = cr->beam();
-    if (b && b->elements().front() == cr) {
-        // beam already considered cross?
-        if (b->cross()) {
-            return false;
-        }
-
-        // for beams not already considered cross,
-        // consider them so here if any elements were moved up
-        for (ChordRest* cr1 : b->elements()) {
-            // some element moved up?
-            if (cr1->staffMove() < 0) {
-                return false;
-            }
-        }
-
-        // not cross
-        return true;
-    }
-
-    // no beam or not first element
-    return false;
-}
-
-//---------------------------------------------------------
 //   notTopBeam
 //    returns true for the first CR of a beam that is cross-staff
 //---------------------------------------------------------
@@ -3367,31 +1932,6 @@ bool notTopBeam(ChordRest* cr)
 }
 
 //---------------------------------------------------------
-//   isTopTuplet
-//    returns true for the first CR of a tuplet that is not cross-staff
-//---------------------------------------------------------
-
-bool isTopTuplet(ChordRest* cr)
-{
-    Tuplet* t = cr->tuplet();
-    if (t && t->elements().front() == cr) {
-        // find top level tuplet
-        while (t->tuplet()) {
-            t = t->tuplet();
-        }
-        // consider tuplet cross if anything moved within it
-        if (t->cross()) {
-            return false;
-        } else {
-            return true;
-        }
-    }
-
-    // no tuplet or not first element
-    return false;
-}
-
-//---------------------------------------------------------
 //   notTopTuplet
 //    returns true for the first CR of a tuplet that is cross-staff
 //---------------------------------------------------------
@@ -3417,1745 +1957,6 @@ bool notTopTuplet(ChordRest* cr)
 }
 
 //---------------------------------------------------------
-//   findLyricsMaxY
-//---------------------------------------------------------
-
-static qreal findLyricsMaxY(Segment& s, int staffIdx)
-{
-    qreal yMax = 0.0;
-    if (!s.isChordRestType()) {
-        return yMax;
-    }
-
-    qreal lyricsMinTopDistance = s.score()->styleP(Sid::lyricsMinTopDistance);
-
-    for (int voice = 0; voice < VOICES; ++voice) {
-        ChordRest* cr = s.cr(staffIdx * VOICES + voice);
-        if (cr && !cr->lyrics().empty()) {
-            SkylineLine sk(true);
-
-            for (Lyrics* l : cr->lyrics()) {
-                if (l->autoplace() && l->placeBelow()) {
-                    qreal yOff = l->offset().y();
-                    PointF offset = l->pos() + cr->pos() + s.pos() + s.measure()->pos();
-                    RectF r = l->bbox().translated(offset);
-                    r.translate(0.0, -yOff);
-                    sk.add(r.x(), r.top(), r.width());
-                }
-            }
-            SysStaff* ss = s.measure()->system()->staff(staffIdx);
-            for (Lyrics* l : cr->lyrics()) {
-                if (l->autoplace() && l->placeBelow()) {
-                    qreal y = ss->skyline().south().minDistance(sk);
-                    if (y > -lyricsMinTopDistance) {
-                        yMax = qMax(yMax, y + lyricsMinTopDistance);
-                    }
-                }
-            }
-        }
-    }
-    return yMax;
-}
-
-//---------------------------------------------------------
-//   findLyricsMinY
-//---------------------------------------------------------
-
-static qreal findLyricsMinY(Segment& s, int staffIdx)
-{
-    qreal yMin = 0.0;
-    if (!s.isChordRestType()) {
-        return yMin;
-    }
-    qreal lyricsMinTopDistance = s.score()->styleP(Sid::lyricsMinTopDistance);
-    for (int voice = 0; voice < VOICES; ++voice) {
-        ChordRest* cr = s.cr(staffIdx * VOICES + voice);
-        if (cr && !cr->lyrics().empty()) {
-            SkylineLine sk(false);
-
-            for (Lyrics* l : cr->lyrics()) {
-                if (l->autoplace() && l->placeAbove()) {
-                    qreal yOff = l->offset().y();
-                    RectF r = l->bbox().translated(l->pos() + cr->pos() + s.pos() + s.measure()->pos());
-                    r.translate(0.0, -yOff);
-                    sk.add(r.x(), r.bottom(), r.width());
-                }
-            }
-            SysStaff* ss = s.measure()->system()->staff(staffIdx);
-            for (Lyrics* l : cr->lyrics()) {
-                if (l->autoplace() && l->placeAbove()) {
-                    qreal y = sk.minDistance(ss->skyline().north());
-                    if (y > -lyricsMinTopDistance) {
-                        yMin = qMin(yMin, -y - lyricsMinTopDistance);
-                    }
-                }
-            }
-        }
-    }
-    return yMin;
-}
-
-static qreal findLyricsMaxY(Measure* m, int staffIdx)
-{
-    qreal yMax = 0.0;
-    for (Segment& s : m->segments()) {
-        yMax = qMax(yMax, findLyricsMaxY(s, staffIdx));
-    }
-    return yMax;
-}
-
-static qreal findLyricsMinY(Measure* m, int staffIdx)
-{
-    qreal yMin = 0.0;
-    for (Segment& s : m->segments()) {
-        yMin = qMin(yMin, findLyricsMinY(s, staffIdx));
-    }
-    return yMin;
-}
-
-//---------------------------------------------------------
-//   applyLyricsMax
-//---------------------------------------------------------
-
-static void applyLyricsMax(Segment& s, int staffIdx, qreal yMax)
-{
-    if (!s.isChordRestType()) {
-        return;
-    }
-    Skyline& sk = s.measure()->system()->staff(staffIdx)->skyline();
-    for (int voice = 0; voice < VOICES; ++voice) {
-        ChordRest* cr = s.cr(staffIdx * VOICES + voice);
-        if (cr && !cr->lyrics().empty()) {
-            qreal lyricsMinBottomDistance = s.score()->styleP(Sid::lyricsMinBottomDistance);
-            for (Lyrics* l : cr->lyrics()) {
-                if (l->autoplace() && l->placeBelow()) {
-                    l->rypos() += yMax - l->propertyDefault(Pid::OFFSET).value<PointF>().y();
-                    if (l->addToSkyline()) {
-                        PointF offset = l->pos() + cr->pos() + s.pos() + s.measure()->pos();
-                        sk.add(l->bbox().translated(offset).adjusted(0.0, 0.0, 0.0, lyricsMinBottomDistance));
-                    }
-                }
-            }
-        }
-    }
-}
-
-static void applyLyricsMax(Measure* m, int staffIdx, qreal yMax)
-{
-    for (Segment& s : m->segments()) {
-        applyLyricsMax(s, staffIdx, yMax);
-    }
-}
-
-//---------------------------------------------------------
-//   applyLyricsMin
-//---------------------------------------------------------
-
-static void applyLyricsMin(ChordRest* cr, int staffIdx, qreal yMin)
-{
-    Skyline& sk = cr->measure()->system()->staff(staffIdx)->skyline();
-    for (Lyrics* l : cr->lyrics()) {
-        if (l->autoplace() && l->placeAbove()) {
-            l->rypos() += yMin - l->propertyDefault(Pid::OFFSET).value<PointF>().y();
-            if (l->addToSkyline()) {
-                PointF offset = l->pos() + cr->pos() + cr->segment()->pos() + cr->segment()->measure()->pos();
-                sk.add(l->bbox().translated(offset));
-            }
-        }
-    }
-}
-
-static void applyLyricsMin(Measure* m, int staffIdx, qreal yMin)
-{
-    for (Segment& s : m->segments()) {
-        if (s.isChordRestType()) {
-            for (int voice = 0; voice < VOICES; ++voice) {
-                ChordRest* cr = s.cr(staffIdx * VOICES + voice);
-                if (cr) {
-                    applyLyricsMin(cr, staffIdx, yMin);
-                }
-            }
-        }
-    }
-}
-
-//---------------------------------------------------------
-//   restoreBeams
-//---------------------------------------------------------
-
-static void restoreBeams(Measure* m)
-{
-    for (Segment* s = m->first(SegmentType::ChordRest); s; s = s->next(SegmentType::ChordRest)) {
-        for (Element* e : s->elist()) {
-            if (e && e->isChordRest()) {
-                ChordRest* cr = toChordRest(e);
-                if (isTopBeam(cr)) {
-                    Beam* b = cr->beam();
-                    b->layout();
-                    b->addSkyline(m->system()->staff(b->staffIdx())->skyline());
-                }
-            }
-        }
-    }
-}
-
-//---------------------------------------------------------
-//   layoutLyrics
-//
-//    vertical align lyrics
-//
-//---------------------------------------------------------
-
-void Score::layoutLyrics(System* system)
-{
-    std::vector<int> visibleStaves;
-    for (int staffIdx = system->firstVisibleStaff(); staffIdx < nstaves(); staffIdx = system->nextVisibleStaff(staffIdx)) {
-        visibleStaves.push_back(staffIdx);
-    }
-
-    //int nAbove[nstaves()];
-    std::vector<int> VnAbove(nstaves());
-
-    for (int staffIdx : visibleStaves) {
-        VnAbove[staffIdx] = 0;
-        for (MeasureBase* mb : system->measures()) {
-            if (!mb->isMeasure()) {
-                continue;
-            }
-            Measure* m = toMeasure(mb);
-            for (Segment& s : m->segments()) {
-                if (s.isChordRestType()) {
-                    for (int voice = 0; voice < VOICES; ++voice) {
-                        ChordRest* cr = s.cr(staffIdx * VOICES + voice);
-                        if (cr) {
-                            int nA = 0;
-                            for (Lyrics* l : cr->lyrics()) {
-                                // user adjusted offset can possibly change placement
-                                if (l->offsetChanged() != OffsetChange::NONE) {
-                                    Placement p = l->placement();
-                                    l->rebaseOffset();
-                                    if (l->placement() != p) {
-                                        l->undoResetProperty(Pid::AUTOPLACE);
-                                        //l->undoResetProperty(Pid::OFFSET);
-                                        //l->layout();
-                                    }
-                                }
-                                l->setOffsetChanged(false);
-                                if (l->placeAbove()) {
-                                    ++nA;
-                                }
-                            }
-                            VnAbove[staffIdx] = qMax(VnAbove[staffIdx], nA);
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    for (int staffIdx : visibleStaves) {
-        for (MeasureBase* mb : system->measures()) {
-            if (!mb->isMeasure()) {
-                continue;
-            }
-            Measure* m = toMeasure(mb);
-            for (Segment& s : m->segments()) {
-                if (s.isChordRestType()) {
-                    for (int voice = 0; voice < VOICES; ++voice) {
-                        ChordRest* cr = s.cr(staffIdx * VOICES + voice);
-                        if (cr) {
-                            for (Lyrics* l : cr->lyrics()) {
-                                l->layout2(VnAbove[staffIdx]);
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    VerticalAlignRange ar = VerticalAlignRange(styleI(Sid::autoplaceVerticalAlignRange));
-
-    switch (ar) {
-    case VerticalAlignRange::MEASURE:
-        for (MeasureBase* mb : system->measures()) {
-            if (!mb->isMeasure()) {
-                continue;
-            }
-            Measure* m = toMeasure(mb);
-            for (int staffIdx : visibleStaves) {
-                qreal yMax = findLyricsMaxY(m, staffIdx);
-                applyLyricsMax(m, staffIdx, yMax);
-            }
-        }
-        break;
-    case VerticalAlignRange::SYSTEM:
-        for (int staffIdx : visibleStaves) {
-            qreal yMax = 0.0;
-            qreal yMin = 0.0;
-            for (MeasureBase* mb : system->measures()) {
-                if (!mb->isMeasure()) {
-                    continue;
-                }
-                yMax = qMax<qreal>(yMax, findLyricsMaxY(toMeasure(mb), staffIdx));
-                yMin = qMin(yMin, findLyricsMinY(toMeasure(mb), staffIdx));
-            }
-            for (MeasureBase* mb : system->measures()) {
-                if (!mb->isMeasure()) {
-                    continue;
-                }
-                applyLyricsMax(toMeasure(mb), staffIdx, yMax);
-                applyLyricsMin(toMeasure(mb), staffIdx, yMin);
-            }
-        }
-        break;
-    case VerticalAlignRange::SEGMENT:
-        for (MeasureBase* mb : system->measures()) {
-            if (!mb->isMeasure()) {
-                continue;
-            }
-            Measure* m = toMeasure(mb);
-            for (int staffIdx : visibleStaves) {
-                for (Segment& s : m->segments()) {
-                    qreal yMax = findLyricsMaxY(s, staffIdx);
-                    applyLyricsMax(s, staffIdx, yMax);
-                }
-            }
-        }
-        break;
-    }
-}
-
-//---------------------------------------------------------
-//   layoutTies
-//---------------------------------------------------------
-
-void layoutTies(Chord* ch, System* system, const Fraction& stick)
-{
-    SysStaff* staff = system->staff(ch->staffIdx());
-    if (!staff->show()) {
-        return;
-    }
-    for (Note* note : ch->notes()) {
-        Tie* t = note->tieFor();
-        if (t) {
-            TieSegment* ts = t->layoutFor(system);
-            if (ts && ts->addToSkyline()) {
-                staff->skyline().add(ts->shape().translated(ts->pos()));
-            }
-        }
-        t = note->tieBack();
-        if (t) {
-            if (t->startNote()->tick() < stick) {
-                TieSegment* ts = t->layoutBack(system);
-                if (ts && ts->addToSkyline()) {
-                    staff->skyline().add(ts->shape().translated(ts->pos()));
-                }
-            }
-        }
-    }
-}
-
-//---------------------------------------------------------
-//   layoutHarmonies
-//---------------------------------------------------------
-
-void layoutHarmonies(const std::vector<Segment*>& sl)
-{
-    for (const Segment* s : sl) {
-        for (Element* e : s->annotations()) {
-            if (e->isHarmony()) {
-                Harmony* h = toHarmony(e);
-                // For chord symbols that coincide with a chord or rest,
-                // a partial layout can also happen (if needed) during ChordRest layout
-                // in order to calculate a bbox and allocate its shape to the ChordRest.
-                // But that layout (if it happens at all) does not do autoplace,
-                // so we need the full layout here.
-                h->layout();
-                h->autoplaceSegmentElement();
-            }
-        }
-    }
-}
-
-//---------------------------------------------------------
-//   alignHarmonies
-//---------------------------------------------------------
-
-void alignHarmonies(const System* system, const std::vector<Segment*>& sl, bool harmony, const qreal maxShiftAbove,
-                    const qreal maxShiftBelow)
-{
-    // Help class.
-    // Contains harmonies/fretboard per segment.
-    class HarmonyList : public QList<Element*>
-    {
-        QMap<const Segment*, QList<Element*> > elements;
-        QList<Element*> modified;
-
-        Element* getReferenceElement(const Segment* s, bool above, bool visible) const
-        {
-            // Returns the reference element for aligning.
-            // When a segments contains multiple harmonies/fretboard, the lowest placed
-            // element (for placement above, otherwise the highest placed element) is
-            // used for alignment.
-            Element* element { nullptr };
-            for (Element* e : elements[s]) {
-                // Only chord symbols have styled offset, fretboards don't.
-                if (!e->autoplace() || (e->isHarmony() && !e->isStyled(Pid::OFFSET)) || (visible && !e->visible())) {
-                    continue;
-                }
-                if (!element) {
-                    element = e;
-                } else {
-                    if ((e->placeAbove() && above && (element->y() < e->y()))
-                        || (e->placeBelow() && !above && (element->y() > e->y()))) {
-                        element = e;
-                    }
-                }
-            }
-            return element;
-        }
-
-    public:
-        HarmonyList()
-        {
-            elements.clear();
-            modified.clear();
-        }
-
-        void append(const Segment* s, Element* e)
-        {
-            elements[s].append(e);
-        }
-
-        qreal getReferenceHeight(bool above) const
-        {
-            // The reference height is the height of
-            //    the lowest element if placed above
-            // or
-            //    the highest element if placed below.
-            bool first { true };
-            qreal ref { 0.0 };
-            for (auto s : elements.keys()) {
-                Element* e { getReferenceElement(s, above, true) };
-                if (!e) {
-                    continue;
-                }
-                if (e->placeAbove() && above) {
-                    ref = first ? e->y() : qMin(ref, e->y());
-                    first = false;
-                } else if (e->placeBelow() && !above) {
-                    ref = first ? e->y() : qMax(ref, e->y());
-                    first = false;
-                }
-            }
-            return ref;
-        }
-
-        bool align(bool above, qreal reference, qreal maxShift)
-        {
-            // Align the elements. If a segment contains multiple elements,
-            // only the reference elements is used in the algorithm. All other
-            // elements will remain their original placement with respect to
-            // the reference element.
-            bool moved { false };
-            if (almostZero(reference)) {
-                return moved;
-            }
-
-            for (auto s : elements.keys()) {
-                QList<Element*> handled;
-                Element* be = getReferenceElement(s, above, false);
-                if (!be) {
-                    // If there are only invisible elements, we have to use an invisible
-                    // element for alignment reference.
-                    be = getReferenceElement(s, above, true);
-                }
-                if (be && ((above && (be->y() < (reference + maxShift))) || ((!above && (be->y() > (reference - maxShift)))))) {
-                    qreal shift = be->rypos();
-                    be->rypos() = reference - be->ryoffset();
-                    shift -= be->rypos();
-                    for (Element* e : elements[s]) {
-                        if ((above && e->placeBelow()) || (!above && e->placeAbove())) {
-                            continue;
-                        }
-                        modified.append(e);
-                        handled.append(e);
-                        moved = true;
-                        if (e != be) {
-                            e->rypos() -= shift;
-                        }
-                    }
-                    for (auto e : handled) {
-                        elements[s].removeOne(e);
-                    }
-                }
-            }
-            return moved;
-        }
-
-        void addToSkyline(const System* system)
-        {
-            for (Element* e : qAsConst(modified)) {
-                const Segment* s = toSegment(e->parent());
-                const MeasureBase* m = toMeasureBase(s->parent());
-                system->staff(e->staffIdx())->skyline().add(e->shape().translated(e->pos() + s->pos() + m->pos()));
-                if (e->isFretDiagram()) {
-                    FretDiagram* fd = toFretDiagram(e);
-                    Harmony* h = fd->harmony();
-                    if (h) {
-                        system->staff(e->staffIdx())->skyline().add(h->shape().translated(h->pos() + fd->pos() + s->pos() + m->pos()));
-                    } else {
-                        system->staff(e->staffIdx())->skyline().add(fd->shape().translated(fd->pos() + s->pos() + m->pos()));
-                    }
-                }
-            }
-        }
-    };
-
-    if (almostZero(maxShiftAbove) && almostZero(maxShiftBelow)) {
-        return;
-    }
-
-    // Collect all fret diagrams and chord symbol and store them per staff.
-    // In the same pass, the maximum height is collected.
-    QMap<int, HarmonyList> staves;
-    for (const Segment* s : sl) {
-        for (Element* e : s->annotations()) {
-            if ((harmony && e->isHarmony()) || (!harmony && e->isFretDiagram())) {
-                staves[e->staffIdx()].append(s, e);
-            }
-        }
-    }
-
-    for (int idx: staves.keys()) {
-        // Align the objects.
-        // Algorithm:
-        //    - Find highest placed harmony/fretdiagram.
-        //    - Align all harmony/fretdiagram objects placed between height and height-maxShiftAbove.
-        //    - Repeat for all harmony/fretdiagram objects below heigt-maxShiftAbove.
-        bool moved { true };
-        int pass { 0 };
-        while (moved && (pass++ < 10)) {
-            moved = false;
-            moved |= staves[idx].align(true, staves[idx].getReferenceHeight(true), maxShiftAbove);
-            moved |= staves[idx].align(false, staves[idx].getReferenceHeight(false), maxShiftBelow);
-        }
-
-        // Add all aligned objects to the sky line.
-        staves[idx].addToSkyline(system);
-    }
-}
-
-//---------------------------------------------------------
-//   processLines
-//---------------------------------------------------------
-
-static void processLines(System* system, std::vector<Spanner*> lines, bool align)
-{
-    std::vector<SpannerSegment*> segments;
-    for (Spanner* sp : lines) {
-        SpannerSegment* ss = sp->layoutSystem(system);         // create/layout spanner segment for this system
-        if (ss->autoplace()) {
-            segments.push_back(ss);
-        }
-    }
-
-    if (align && segments.size() > 1) {
-        const int nstaves = system->staves()->size();
-        constexpr qreal minY = -1000000.0;
-        const qreal defaultY = segments[0]->rypos();
-        std::vector<qreal> y(nstaves, minY);
-
-        for (SpannerSegment* ss : segments) {
-            if (ss->visible()) {
-                qreal& staffY = y[ss->staffIdx()];
-                staffY = qMax(staffY, ss->rypos());
-            }
-        }
-        for (SpannerSegment* ss : segments) {
-            if (!ss->isStyled(Pid::OFFSET)) {
-                continue;
-            }
-            const qreal staffY = y[ss->staffIdx()];
-            if (staffY > minY) {
-                ss->rypos() = staffY;
-            } else {
-                ss->rypos() = defaultY;
-            }
-        }
-    }
-
-    //
-    // add shapes to skyline
-    //
-    for (SpannerSegment* ss : segments) {
-        if (ss->addToSkyline()) {
-            system->staff(ss->staffIdx())->skyline().add(ss->shape().translated(ss->pos()));
-        }
-    }
-}
-
-//---------------------------------------------------------
-//   collectSystem
-//---------------------------------------------------------
-
-System* Score::collectSystem(LayoutContext& lc)
-{
-    if (!lc.curMeasure) {
-        return 0;
-    }
-    const MeasureBase* measure  = _systems.empty() ? 0 : _systems.back()->measures().back();
-    if (measure) {
-        measure = measure->findPotentialSectionBreak();
-    }
-    if (measure) {
-        lc.firstSystem        = measure->sectionBreak() && _layoutMode != LayoutMode::FLOAT;
-        lc.firstSystemIndent  = lc.firstSystem && measure->sectionBreakElement()->firstSystemIdentation() && styleB(
-            Sid::enableIndentationOnFirstSystem);
-        lc.startWithLongNames = lc.firstSystem && measure->sectionBreakElement()->startWithLongNames();
-    }
-    System* system = getNextSystem(lc);
-    Fraction lcmTick = lc.curMeasure->tick();
-    system->setInstrumentNames(lc.startWithLongNames, lcmTick);
-
-    qreal minWidth    = 0;
-    qreal layoutSystemMinWidth = 0;
-    bool firstMeasure = true;
-    bool createHeader = false;
-    qreal systemWidth = styleD(Sid::pagePrintableWidth) * DPI;
-    system->setWidth(systemWidth);
-
-    // save state of measure
-    bool curHeader = lc.curMeasure->header();
-    bool curTrailer = lc.curMeasure->trailer();
-    MeasureBase* breakMeasure = nullptr;
-
-    while (lc.curMeasure) {      // collect measure for system
-        System* oldSystem = lc.curMeasure->system();
-        system->appendMeasure(lc.curMeasure);
-
-        qreal ww  = 0;          // width of current measure
-
-        if (lc.curMeasure->isMeasure()) {
-            Measure* m = toMeasure(lc.curMeasure);
-            if (firstMeasure) {
-                layoutSystemMinWidth = minWidth;
-                system->layoutSystem(minWidth, lc.firstSystem, lc.firstSystemIndent);
-                minWidth += system->leftMargin();
-                if (m->repeatStart()) {
-                    Segment* s = m->findSegmentR(SegmentType::StartRepeatBarLine, Fraction(0, 1));
-                    if (!s->enabled()) {
-                        s->setEnabled(true);
-                    }
-                }
-                m->addSystemHeader(lc.firstSystem);
-                firstMeasure = false;
-                createHeader = false;
-            } else {
-                if (createHeader) {
-                    m->addSystemHeader(false);
-                    createHeader = false;
-                } else if (m->header()) {
-                    m->removeSystemHeader();
-                }
-            }
-
-            m->createEndBarLines(true);
-            // measures with nobreak cannot end a system
-            // thus they will not contain a trailer
-            if (m->noBreak()) {
-                m->removeSystemTrailer();
-            } else {
-                m->addSystemTrailer(m->nextMeasure());
-            }
-            m->computeMinWidth();
-            ww = m->width();
-        } else if (lc.curMeasure->isHBox()) {
-            lc.curMeasure->computeMinWidth();
-            ww = lc.curMeasure->width();
-            createHeader = toHBox(lc.curMeasure)->createSystemHeader();
-        } else {
-            // vbox:
-            getNextMeasure(lc);
-            system->layout2();         // compute staff distances
-            return system;
-        }
-        // check if lc.curMeasure fits, remove if not
-        // collect at least one measure and the break
-
-        bool doBreak = (system->measures().size() > 1) && ((minWidth + ww) > systemWidth);
-        if (doBreak) {
-            breakMeasure = lc.curMeasure;
-            system->removeLastMeasure();
-            lc.curMeasure->setSystem(oldSystem);
-            while (lc.prevMeasure && lc.prevMeasure->noBreak() && system->measures().size() > 1) {
-                // remove however many measures are grouped with nobreak, working backwards
-                // but if too many are grouped, stop before we get 0 measures left on system
-                // TODO: intelligently break group into smaller groups instead
-                lc.tick -= lc.curMeasure->ticks();
-                lc.measureNo = lc.prevMeasure->no();
-
-                lc.nextMeasure = lc.curMeasure;
-                lc.curMeasure  = lc.prevMeasure;
-                lc.prevMeasure = lc.curMeasure->prevMeasure();
-
-                minWidth -= system->lastMeasure()->width();
-                system->removeLastMeasure();
-                lc.curMeasure->setSystem(oldSystem);
-            }
-            break;
-        }
-
-        if (lc.prevMeasure && lc.prevMeasure->isMeasure() && lc.prevMeasure->system() == system) {
-            //
-            // now we know that the previous measure is not the last
-            // measure in the system and we finally can create the end barline for it
-
-            Measure* m = toMeasure(lc.prevMeasure);
-            // TODO: if lc.curMeasure is a frame, removing the trailer may be premature
-            // but merely skipping this code isn't good enough,
-            // we need to find the right time to re-enable the trailer,
-            // since it seems to be disabled somewhere else
-            if (m->trailer()) {
-                qreal ow = m->width();
-                m->removeSystemTrailer();
-                minWidth += m->width() - ow;
-            }
-            // if the prev measure is an end repeat and the cur measure
-            // is an repeat, the createEndBarLines() created an start-end repeat barline
-            // and we can remove the start repeat barline of the current barline
-
-            if (lc.curMeasure->isMeasure()) {
-                Measure* m1 = toMeasure(lc.curMeasure);
-                if (m1->repeatStart()) {
-                    Segment* s = m1->findSegmentR(SegmentType::StartRepeatBarLine, Fraction(0, 1));
-                    if (!s->enabled()) {
-                        s->setEnabled(true);
-                        m1->computeMinWidth();
-                        ww = m1->width();
-                    }
-                }
-            }
-            // TODO: we actually still don't know for sure
-            // if this will be the last true measure of the system or not
-            // since the lc.curMeasure may be a frame
-            // but at this point we have no choice but to assume it isn't
-            // since we don't know yet if another true measure will fit
-            // worst that happens is we don't get the automatic double bar before a courtesy key signature
-            minWidth += m->createEndBarLines(false);          // create final barLine
-        }
-
-        MeasureBase* mb = lc.curMeasure;
-        bool lineBreak  = false;
-        switch (_layoutMode) {
-        case LayoutMode::PAGE:
-        case LayoutMode::SYSTEM:
-            lineBreak = mb->pageBreak() || mb->lineBreak() || mb->sectionBreak();
-            break;
-        case LayoutMode::FLOAT:
-        case LayoutMode::LINE:
-            lineBreak = false;
-            break;
-        }
-
-        // preserve state of next measure (which is about to become current measure)
-        if (lc.nextMeasure) {
-            MeasureBase* nmb = lc.nextMeasure;
-            if (nmb->isMeasure() && styleB(Sid::createMultiMeasureRests)) {
-                Measure* nm = toMeasure(nmb);
-                if (nm->hasMMRest()) {
-                    nmb = nm->mmRest();
-                }
-            }
-            nmb->setOldWidth(nmb->width());
-            if (!lc.curMeasure->noBreak()) {
-                // current measure is not a nobreak,
-                // so next measure could possibly start a system
-                curHeader = nmb->header();
-            }
-            if (!nmb->noBreak()) {
-                // next measure is not a nobreak
-                // so it could possibly end a system
-                curTrailer = nmb->trailer();
-            }
-        }
-
-        getNextMeasure(lc);
-
-        minWidth += ww;
-
-        // ElementType nt = lc.curMeasure ? lc.curMeasure->type() : ElementType::INVALID;
-        mb = lc.curMeasure;
-        bool tooWide = false;     // minWidth + minMeasureWidth > systemWidth;  // TODO: noBreak
-        if (lineBreak || !mb || mb->isVBox() || mb->isTBox() || mb->isFBox() || tooWide) {
-            break;
-        }
-    }
-
-    if (lc.endTick < lc.prevMeasure->tick()) {
-        // we've processed the entire range
-        // but we need to continue layout until we reach a system whose last measure is the same as previous layout
-        if (lc.prevMeasure == lc.systemOldMeasure) {
-            // this system ends in the same place as the previous layout
-            // ok to stop
-            if (lc.curMeasure && lc.curMeasure->isMeasure()) {
-                // we may have previously processed first measure(s) of next system
-                // so now we must restore to original state
-                Measure* m = toMeasure(lc.curMeasure);
-                if (m->repeatStart()) {
-                    Segment* s = m->findSegmentR(SegmentType::StartRepeatBarLine, Fraction(0, 1));
-                    if (!s->enabled()) {
-                        s->setEnabled(true);
-                    }
-                }
-                const MeasureBase* pbmb = lc.prevMeasure->findPotentialSectionBreak();
-                bool firstSystem = pbmb->sectionBreak() && _layoutMode != LayoutMode::FLOAT;
-                MeasureBase* nm = breakMeasure ? breakMeasure : m;
-                if (curHeader) {
-                    m->addSystemHeader(firstSystem);
-                } else {
-                    m->removeSystemHeader();
-                }
-                for (;;) {
-                    // TODO: what if the nobreak group takes the entire system - is this correct?
-                    if (curTrailer && !m->noBreak()) {
-                        m->addSystemTrailer(m->nextMeasure());
-                    } else {
-                        m->removeSystemTrailer();
-                    }
-                    m->computeMinWidth();
-                    m->stretchMeasure(m->oldWidth());
-                    restoreBeams(m);
-                    if (m == nm || !m->noBreak()) {
-                        break;
-                    }
-                    m = m->nextMeasure();
-                }
-            }
-            lc.rangeDone = true;
-        }
-    }
-
-    //
-    // now we have a complete set of measures for this system
-    //
-    // prevMeasure is the last measure in the system
-    if (lc.prevMeasure && lc.prevMeasure->isMeasure()) {
-        breakCrossMeasureBeams(toMeasure(lc.prevMeasure));
-        qreal w = toMeasure(lc.prevMeasure)->createEndBarLines(true);
-        minWidth += w;
-    }
-
-    hideEmptyStaves(system, lc.firstSystem);
-    // Relayout system decorations to reuse space properly for
-    // hidden staves' instrument names or other hidden elements.
-    minWidth -= system->leftMargin();
-    system->layoutSystem(layoutSystemMinWidth, lc.firstSystem, lc.firstSystemIndent);
-    minWidth += system->leftMargin();
-
-    //-------------------------------------------------------
-    //    add system trailer if needed
-    //    (cautionary time/key signatures etc)
-    //-------------------------------------------------------
-
-    Measure* lm  = system->lastMeasure();
-    if (lm) {
-        Measure* nm = lm->nextMeasure();
-        if (nm) {
-            qreal w = lm->width();
-            lm->addSystemTrailer(nm);
-            if (lm->trailer()) {
-                lm->computeMinWidth();
-            }
-            minWidth += lm->width() - w;
-        }
-    }
-
-    //
-    // stretch incomplete row
-    //
-    qreal rest;
-    if (MScore::noHorizontalStretch) {
-        rest = 0;
-    } else {
-        qreal mw          = system->leftMargin();          // DEBUG
-        qreal totalWeight = 0.0;
-
-        for (MeasureBase* mb : system->measures()) {
-            if (mb->isHBox()) {
-                mw += mb->width();
-            } else if (mb->isMeasure()) {
-                Measure* m  = toMeasure(mb);
-                mw          += m->width();                       // measures are stretched already with basicStretch()
-                int weight   = m->layoutWeight();
-                totalWeight += weight * m->basicStretch();
-            }
-        }
-
-#ifndef NDEBUG
-        if (!qFuzzyCompare(mw, minWidth)) {
-            qDebug("==layoutSystem %6d old %.1f new %.1f", system->measures().front()->tick().ticks(), minWidth, mw);
-        }
-#endif
-        rest = systemWidth - minWidth;
-        //
-        // don’t stretch last system row, if accumulated minWidth is <= lastSystemFillLimit
-        //
-        if (lc.curMeasure == 0 && ((minWidth / systemWidth) <= styleD(Sid::lastSystemFillLimit))) {
-            if (minWidth > rest) {
-                rest = rest * .5;
-            } else {
-                rest = minWidth;
-            }
-        }
-        rest /= totalWeight;
-    }
-
-    PointF pos;
-    firstMeasure = true;
-    bool createBrackets = false;
-    for (MeasureBase* mb : system->measures()) {
-        qreal ww = mb->width();
-        if (mb->isMeasure()) {
-            if (firstMeasure) {
-                pos.rx() += system->leftMargin();
-                firstMeasure = false;
-            }
-            mb->setPos(pos);
-            Measure* m = toMeasure(mb);
-            qreal stretch = m->basicStretch();
-            int weight = m->layoutWeight();
-            ww  += rest * weight * stretch;
-            m->stretchMeasure(ww);
-            m->layoutStaffLines();
-            if (createBrackets) {
-                system->addBrackets(toMeasure(mb));
-                createBrackets = false;
-            }
-        } else if (mb->isHBox()) {
-            mb->setPos(pos + PointF(toHBox(mb)->topGap(), 0.0));
-            mb->layout();
-            createBrackets = toHBox(mb)->createSystemHeader();
-        } else if (mb->isVBox()) {
-            mb->setPos(pos);
-        }
-        pos.rx() += ww;
-    }
-    system->setWidth(pos.x());
-
-    layoutSystemElements(system, lc);
-    system->layout2();     // compute staff distances
-    // TODO: now that the code at the top of this function does this same backwards search,
-    // we might be able to eliminate this block
-    // but, lc might be used elsewhere so we need to be careful
-#if 1
-    measure = system->measures().back();
-    if (measure) {
-        measure = measure->findPotentialSectionBreak();
-    }
-    if (measure) {
-        lc.firstSystem        = measure->sectionBreak() && _layoutMode != LayoutMode::FLOAT;
-        lc.firstSystemIndent  = lc.firstSystem && measure->sectionBreakElement()->firstSystemIdentation() && styleB(
-            Sid::enableIndentationOnFirstSystem);
-        lc.startWithLongNames = lc.firstSystem && measure->sectionBreakElement()->startWithLongNames();
-    }
-#endif
-    return system;
-}
-
-//---------------------------------------------------------
-//   layoutSystemElements
-//---------------------------------------------------------
-
-void Score::layoutSystemElements(System* system, LayoutContext& lc)
-{
-    //-------------------------------------------------------------
-    //    create cr segment list to speed up computations
-    //-------------------------------------------------------------
-
-    std::vector<Segment*> sl;
-    for (MeasureBase* mb : system->measures()) {
-        if (!mb->isMeasure()) {
-            continue;
-        }
-        Measure* m = toMeasure(mb);
-        m->layoutMeasureNumber();
-        m->layoutMMRestRange();
-
-        // in continuous view, entire score is one system
-        // but we only need to process the range
-        if (lineMode() && (m->tick() < lc.startTick || m->tick() > lc.endTick)) {
-            continue;
-        }
-        for (Segment* s = m->first(); s; s = s->next()) {
-            if (s->isChordRestType() || !s->annotations().empty()) {
-                sl.push_back(s);
-            }
-        }
-    }
-
-    //-------------------------------------------------------------
-    // layout beams
-    //  Needs to be done before creating skylines as stem lengths
-    //  may change.
-    //-------------------------------------------------------------
-
-    for (Segment* s : sl) {
-        for (Element* e : s->elist()) {
-            if (!e || !e->isChordRest() || !score()->staff(e->staffIdx())->show()) {
-                // the beam and its system may still be referenced when selecting all,
-                // even if the staff is invisible. The old system is invalid and does cause problems in #284012
-                if (e && e->isChordRest() && !score()->staff(e->staffIdx())->show() && toChordRest(e)->beam()) {
-                    toChordRest(e)->beam()->setParent(nullptr);
-                }
-                continue;
-            }
-            ChordRest* cr = toChordRest(e);
-
-            // layout beam
-            if (isTopBeam(cr)) {
-                Beam* b = cr->beam();
-                b->layout();
-            }
-        }
-    }
-
-    //-------------------------------------------------------------
-    //    create skylines
-    //-------------------------------------------------------------
-
-    for (int staffIdx = 0; staffIdx < nstaves(); ++staffIdx) {
-        SysStaff* ss = system->staff(staffIdx);
-        Skyline& skyline = ss->skyline();
-        skyline.clear();
-        for (MeasureBase* mb : system->measures()) {
-            if (!mb->isMeasure()) {
-                continue;
-            }
-            Measure* m = toMeasure(mb);
-            MeasureNumber* mno = m->noText(staffIdx);
-            MMRestRange* mmrr  = m->mmRangeText(staffIdx);
-            // no need to build skyline outside of range in continuous view
-            if (lineMode() && (m->tick() < lc.startTick || m->tick() > lc.endTick)) {
-                continue;
-            }
-            if (mno && mno->addToSkyline()) {
-                ss->skyline().add(mno->bbox().translated(m->pos() + mno->pos()));
-            }
-            if (mmrr && mmrr->addToSkyline()) {
-                ss->skyline().add(mmrr->bbox().translated(m->pos() + mmrr->pos()));
-            }
-            if (m->staffLines(staffIdx)->addToSkyline()) {
-                ss->skyline().add(m->staffLines(staffIdx)->bbox().translated(m->pos()));
-            }
-            for (Segment& s : m->segments()) {
-                if (!s.enabled() || s.isTimeSigType()) {             // hack: ignore time signatures
-                    continue;
-                }
-                PointF p(s.pos() + m->pos());
-                if (s.segmentType()
-                    & (SegmentType::BarLine | SegmentType::EndBarLine | SegmentType::StartRepeatBarLine | SegmentType::BeginBarLine)) {
-                    BarLine* bl = toBarLine(s.element(staffIdx * VOICES));
-                    if (bl && bl->addToSkyline()) {
-                        RectF r = bl->layoutRect();
-                        skyline.add(r.translated(bl->pos() + p));
-                    }
-                } else {
-                    int strack = staffIdx * VOICES;
-                    int etrack = strack + VOICES;
-                    for (Element* e : s.elist()) {
-                        if (!e) {
-                            continue;
-                        }
-                        int effectiveTrack = e->vStaffIdx() * VOICES + e->voice();
-                        if (effectiveTrack < strack || effectiveTrack >= etrack) {
-                            continue;
-                        }
-
-                        // clear layout for chord-based fingerings
-                        // do this before adding chord to skyline
-                        if (e->isChord()) {
-                            Chord* c = toChord(e);
-                            std::list<Note*> notes;
-                            for (auto gc : c->graceNotes()) {
-                                for (auto n : gc->notes()) {
-                                    notes.push_back(n);
-                                }
-                            }
-                            for (auto n : c->notes()) {
-                                notes.push_back(n);
-                            }
-                            for (Note* note : notes) {
-                                for (Element* en : note->el()) {
-                                    if (en->isFingering()) {
-                                        Fingering* f = toFingering(en);
-                                        if (f->layoutType() == ElementType::CHORD) {
-                                            f->setPos(PointF());
-                                            f->setbbox(RectF());
-                                        }
-                                    }
-                                }
-                            }
-                        }
-
-                        // add element to skyline
-                        if (e->addToSkyline()) {
-                            skyline.add(e->shape().translated(e->pos() + p));
-                        }
-
-                        // add tremolo to skyline
-                        if (e->isChord() && toChord(e)->tremolo()) {
-                            Tremolo* t = toChord(e)->tremolo();
-                            Chord* c1 = t->chord1();
-                            Chord* c2 = t->chord2();
-                            if (!t->twoNotes() || (c1 && !c1->staffMove() && c2 && !c2->staffMove())) {
-                                if (t->chord() == e && t->addToSkyline()) {
-                                    skyline.add(t->shape().translated(t->pos() + e->pos() + p));
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    //-------------------------------------------------------------
-    // layout fingerings, add beams to skylines
-    //-------------------------------------------------------------
-
-    for (Segment* s : sl) {
-        std::set<int> recreateShapes;
-        for (Element* e : s->elist()) {
-            if (!e || !e->isChordRest() || !score()->staff(e->staffIdx())->show()) {
-                continue;
-            }
-            ChordRest* cr = toChordRest(e);
-
-            // add beam to skyline
-            if (isTopBeam(cr)) {
-                Beam* b = cr->beam();
-                b->addSkyline(system->staff(b->staffIdx())->skyline());
-            }
-
-            // layout chord-based fingerings
-            if (e->isChord()) {
-                Chord* c = toChord(e);
-                std::list<Note*> notes;
-                for (auto gc : c->graceNotes()) {
-                    for (auto n : gc->notes()) {
-                        notes.push_back(n);
-                    }
-                }
-                for (auto n : c->notes()) {
-                    notes.push_back(n);
-                }
-                std::list<Fingering*> fingerings;
-                for (Note* note : notes) {
-                    for (Element* el : note->el()) {
-                        if (el->isFingering()) {
-                            Fingering* f = toFingering(el);
-                            if (f->layoutType() == ElementType::CHORD) {
-                                if (f->placeAbove()) {
-                                    fingerings.push_back(f);
-                                } else {
-                                    fingerings.push_front(f);
-                                }
-                            }
-                        }
-                    }
-                }
-                for (Fingering* f : fingerings) {
-                    f->layout();
-                    if (f->addToSkyline()) {
-                        Note* n = f->note();
-                        RectF r = f->bbox().translated(f->pos() + n->pos() + n->chord()->pos() + s->pos() + s->measure()->pos());
-                        system->staff(f->note()->chord()->vStaffIdx())->skyline().add(r);
-                    }
-                    recreateShapes.insert(f->staffIdx());
-                }
-            }
-        }
-        for (auto staffIdx : recreateShapes) {
-            s->createShape(staffIdx);
-        }
-    }
-
-    //-------------------------------------------------------------
-    // layout articulations
-    //-------------------------------------------------------------
-
-    for (Segment* s : sl) {
-        for (Element* e : s->elist()) {
-            if (!e || !e->isChordRest() || !score()->staff(e->staffIdx())->show()) {
-                continue;
-            }
-            ChordRest* cr = toChordRest(e);
-            // articulations
-            if (cr->isChord()) {
-                Chord* c = toChord(cr);
-                c->layoutArticulations();
-                c->layoutArticulations2();
-            }
-        }
-    }
-
-    //-------------------------------------------------------------
-    // layout tuplets
-    //-------------------------------------------------------------
-
-    for (Segment* s : sl) {
-        for (Element* e : s->elist()) {
-            if (!e || !e->isChordRest() || !score()->staff(e->staffIdx())->show()) {
-                continue;
-            }
-            ChordRest* cr = toChordRest(e);
-            if (!isTopTuplet(cr)) {
-                continue;
-            }
-            DurationElement* de = cr;
-            while (de->tuplet() && de->tuplet()->elements().front() == de) {
-                Tuplet* t = de->tuplet();
-                t->layout();
-                de = t;
-            }
-        }
-    }
-
-    //-------------------------------------------------------------
-    // Drumline sticking
-    //-------------------------------------------------------------
-
-    for (const Segment* s : sl) {
-        for (Element* e : s->annotations()) {
-            if (e->isSticking()) {
-                e->layout();
-            }
-        }
-    }
-
-    //-------------------------------------------------------------
-    // layout slurs
-    //-------------------------------------------------------------
-
-    bool useRange = false;    // TODO: lineMode();
-    Fraction stick = useRange ? lc.startTick : system->measures().front()->tick();
-    Fraction etick = useRange ? lc.endTick : system->measures().back()->endTick();
-    auto spanners = score()->spannerMap().findOverlapping(stick.ticks(), etick.ticks());
-
-    std::vector<Spanner*> spanner;
-    for (auto interval : spanners) {
-        Spanner* sp = interval.value;
-        sp->computeStartElement();
-        sp->computeEndElement();
-        lc.processedSpanners.insert(sp);
-        if (sp->tick() < etick && sp->tick2() >= stick) {
-            if (sp->isSlur()) {
-                // skip cross-staff slurs
-                ChordRest* scr = sp->startCR();
-                ChordRest* ecr = sp->endCR();
-                int idx = sp->vStaffIdx();
-                if (scr && ecr && (scr->vStaffIdx() != idx || ecr->vStaffIdx() != idx)) {
-                    continue;
-                }
-                spanner.push_back(sp);
-            }
-        }
-    }
-    processLines(system, spanner, false);
-    for (auto s : spanner) {
-        Slur* slur = toSlur(s);
-        ChordRest* scr = s->startCR();
-        ChordRest* ecr = s->endCR();
-        if (scr && scr->isChord()) {
-            toChord(scr)->layoutArticulations3(slur);
-        }
-        if (ecr && ecr->isChord()) {
-            toChord(ecr)->layoutArticulations3(slur);
-        }
-    }
-
-    std::vector<Dynamic*> dynamics;
-    for (Segment* s : sl) {
-        for (Element* e : s->elist()) {
-            if (!e) {
-                continue;
-            }
-            if (e->isChord()) {
-                Chord* c = toChord(e);
-                for (Chord* ch : c->graceNotes()) {
-                    layoutTies(ch, system, stick);
-                }
-                layoutTies(c, system, stick);
-            }
-        }
-        for (Element* e : s->annotations()) {
-            if (e->isDynamic()) {
-                Dynamic* d = toDynamic(e);
-                d->layout();
-
-                if (d->autoplace()) {
-                    d->autoplaceSegmentElement(false);
-                    dynamics.push_back(d);
-                }
-            } else if (e->isFiguredBass()) {
-                e->layout();
-                e->autoplaceSegmentElement();
-            }
-        }
-    }
-
-    // add dynamics shape to skyline
-
-    for (Dynamic* d : dynamics) {
-        if (!d->addToSkyline()) {
-            continue;
-        }
-        int si = d->staffIdx();
-        Segment* s = d->segment();
-        Measure* m = s->measure();
-        system->staff(si)->skyline().add(d->shape().translated(d->pos() + s->pos() + m->pos()));
-    }
-
-    //-------------------------------------------------------------
-    // layout SpannerSegments for current system
-    // ottavas, pedals, voltas are collected here, but layouted later
-    //-------------------------------------------------------------
-
-    spanner.clear();
-    std::vector<Spanner*> hairpins;
-    std::vector<Spanner*> ottavas;
-    std::vector<Spanner*> pedal;
-    std::vector<Spanner*> voltas;
-
-    for (auto interval : spanners) {
-        Spanner* sp = interval.value;
-        if (sp->tick() < etick && sp->tick2() > stick) {
-            if (sp->isOttava()) {
-                ottavas.push_back(sp);
-            } else if (sp->isPedal()) {
-                pedal.push_back(sp);
-            } else if (sp->isVolta()) {
-                voltas.push_back(sp);
-            } else if (sp->isHairpin()) {
-                hairpins.push_back(sp);
-            } else if (!sp->isSlur() && !sp->isVolta()) {      // slurs are already
-                spanner.push_back(sp);
-            }
-        }
-    }
-    processLines(system, hairpins, false);
-    processLines(system, spanner, false);
-
-    //-------------------------------------------------------------
-    // Fermata, TremoloBar
-    //-------------------------------------------------------------
-
-    for (const Segment* s : sl) {
-        for (Element* e : s->annotations()) {
-            if (e->isFermata() || e->isTremoloBar()) {
-                e->layout();
-            }
-        }
-    }
-
-    //-------------------------------------------------------------
-    // Ottava, Pedal
-    //-------------------------------------------------------------
-
-    processLines(system, ottavas, false);
-    processLines(system, pedal,   true);
-
-    //-------------------------------------------------------------
-    // Lyric
-    //-------------------------------------------------------------
-
-    layoutLyrics(system);
-
-    // here are lyrics dashes and melisma
-    for (Spanner* sp : _unmanagedSpanner) {
-        if (sp->tick() >= etick || sp->tick2() <= stick) {
-            continue;
-        }
-        sp->layoutSystem(system);
-    }
-
-    //
-    // We need to known if we have FretDiagrams in the system to decide when to layout the Harmonies
-    //
-
-    bool hasFretDiagram = false;
-    for (const Segment* s : sl) {
-        for (Element* e : s->annotations()) {
-            if (e->isFretDiagram()) {
-                hasFretDiagram = true;
-                break;
-            }
-        }
-
-        if (hasFretDiagram) {
-            break;
-        }
-    }
-
-    //-------------------------------------------------------------
-    // Harmony, 1st place
-    // If we have FretDiagrams, we want the Harmony above this and
-    // above the volta, therefore we delay the layout.
-    //-------------------------------------------------------------
-
-    if (!hasFretDiagram) {
-        layoutHarmonies(sl);
-        alignHarmonies(system, sl, true, styleP(Sid::maxChordShiftAbove), styleP(Sid::maxChordShiftBelow));
-    }
-
-    //-------------------------------------------------------------
-    // StaffText, InstrumentChange
-    //-------------------------------------------------------------
-
-    for (const Segment* s : sl) {
-        for (Element* e : s->annotations()) {
-            if (e->isStaffText() || e->isSystemText() || e->isInstrumentChange()) {
-                e->layout();
-            }
-        }
-    }
-
-    //-------------------------------------------------------------
-    // Jump, Marker
-    //-------------------------------------------------------------
-
-    for (MeasureBase* mb : system->measures()) {
-        if (!mb->isMeasure()) {
-            continue;
-        }
-        Measure* m = toMeasure(mb);
-        for (Element* e : m->el()) {
-            if (e->isJump() || e->isMarker()) {
-                e->layout();
-            }
-        }
-    }
-
-    //-------------------------------------------------------------
-    // TempoText
-    //-------------------------------------------------------------
-
-    for (const Segment* s : sl) {
-        for (Element* e : s->annotations()) {
-            if (e->isTempoText()) {
-                e->layout();
-            }
-        }
-    }
-
-    //-------------------------------------------------------------
-    // layout Voltas for current system
-    //-------------------------------------------------------------
-
-    processLines(system, voltas, false);
-
-    //
-    // vertical align volta segments
-    //
-    for (int staffIdx = 0; staffIdx < nstaves(); ++staffIdx) {
-        std::vector<SpannerSegment*> voltaSegments;
-        for (SpannerSegment* ss : system->spannerSegments()) {
-            if (ss->isVoltaSegment() && ss->staffIdx() == staffIdx) {
-                voltaSegments.push_back(ss);
-            }
-        }
-        while (!voltaSegments.empty()) {
-            // we assume voltas are sorted left to right (by tick values)
-            qreal y = 0;
-            int idx = 0;
-            Volta* prevVolta = 0;
-            for (SpannerSegment* ss : voltaSegments) {
-                Volta* volta = toVolta(ss->spanner());
-                if (prevVolta && prevVolta != volta) {
-                    // check if volta is adjacent to prevVolta
-                    if (prevVolta->tick2() != volta->tick()) {
-                        break;
-                    }
-                }
-                y = qMin(y, ss->rypos());
-                ++idx;
-                prevVolta = volta;
-            }
-
-            for (int i = 0; i < idx; ++i) {
-                SpannerSegment* ss = voltaSegments[i];
-                if (ss->autoplace() && ss->isStyled(Pid::OFFSET)) {
-                    ss->rypos() = y;
-                }
-                if (ss->addToSkyline()) {
-                    system->staff(staffIdx)->skyline().add(ss->shape().translated(ss->pos()));
-                }
-            }
-
-            voltaSegments.erase(voltaSegments.begin(), voltaSegments.begin() + idx);
-        }
-    }
-
-    //-------------------------------------------------------------
-    // FretDiagram
-    //-------------------------------------------------------------
-
-    if (hasFretDiagram) {
-        for (const Segment* s : sl) {
-            for (Element* e : s->annotations()) {
-                if (e->isFretDiagram()) {
-                    e->layout();
-                }
-            }
-        }
-
-        //-------------------------------------------------------------
-        // Harmony, 2nd place
-        // We have FretDiagrams, we want the Harmony above this and
-        // above the volta.
-        //-------------------------------------------------------------
-
-        layoutHarmonies(sl);
-        alignHarmonies(system, sl, false, styleP(Sid::maxFretShiftAbove), styleP(Sid::maxFretShiftBelow));
-    }
-
-    //-------------------------------------------------------------
-    // RehearsalMark
-    //-------------------------------------------------------------
-
-    for (const Segment* s : sl) {
-        for (Element* e : s->annotations()) {
-            if (e->isRehearsalMark()) {
-                e->layout();
-            }
-        }
-    }
-
-    //-------------------------------------------------------------
-    // Image
-    //-------------------------------------------------------------
-
-    for (const Segment* s : sl) {
-        for (Element* e : s->annotations()) {
-            if (e->isImage()) {
-                e->layout();
-            }
-        }
-    }
-}
-
-//---------------------------------------------------------
-//   collectPage
-//---------------------------------------------------------
-
-void LayoutContext::collectPage()
-{
-    const qreal slb = score->styleP(Sid::staffLowerBorder);
-    bool breakPages = score->layoutMode() != LayoutMode::SYSTEM;
-    qreal ey        = page->height() - page->bm();
-    qreal y         = 0.0;
-
-    System* nextSystem = 0;
-    int systemIdx = -1;
-
-    // re-calculate positions for systems before current
-    // (they may have been filled on previous layout)
-    int pSystems = page->systems().size();
-    if (pSystems > 0) {
-        page->system(0)->restoreLayout2();
-        y = page->system(0)->y() + page->system(0)->height();
-    } else {
-        y = page->tm();
-    }
-    for (int i = 1; i < pSystems; ++i) {
-        System* cs = page->system(i);
-        System* ps = page->system(i - 1);
-        qreal distance = ps->minDistance(cs);
-        y += distance;
-        cs->setPos(page->lm(), y);
-        cs->restoreLayout2();
-        y += cs->height();
-    }
-
-    for (int k = 0;; ++k) {
-        //
-        // calculate distance to previous system
-        //
-        qreal distance;
-        if (prevSystem) {
-            distance = prevSystem->minDistance(curSystem);
-        } else {
-            // this is the first system on page
-            if (curSystem->vbox()) {
-                distance = 0.0;
-            } else {
-                distance = score->styleP(Sid::staffUpperBorder);
-                bool fixedDistance = false;
-                // TODO: curSystem->spacerDistance(true)
-                for (MeasureBase* mb : curSystem->measures()) {
-                    if (mb->isMeasure()) {
-                        Measure* m = toMeasure(mb);
-                        Spacer* sp = m->vspacerUp(0);                   // TODO: first visible?
-                        if (sp) {
-                            if (sp->spacerType() == SpacerType::FIXED) {
-                                distance = sp->gap();
-                                fixedDistance = true;
-                                break;
-                            } else {
-                                distance = qMax(distance, sp->gap());
-                            }
-                        }
-//TODO::ws                                    distance = qMax(distance, -m->staffShape(0).top());
-                    }
-                }
-                if (!fixedDistance) {
-                    distance = qMax(distance, curSystem->minTop());
-                }
-            }
-        }
-//TODO-ws ??
-//          distance += score->staves().front()->userDist();
-
-        y += distance;
-        curSystem->setPos(page->lm(), y);
-        curSystem->restoreLayout2();
-        page->appendSystem(curSystem);
-        y += curSystem->height();
-
-        //
-        //  check for page break or if next system will fit on page
-        //
-        bool collected = false;
-        if (rangeDone) {
-            // take next system unchanged
-            if (systemIdx > 0) {
-                nextSystem = score->systems().value(systemIdx++);
-                if (!nextSystem) {
-                    // TODO: handle next movement
-                }
-            } else {
-                nextSystem = systemList.empty() ? 0 : systemList.takeFirst();
-                if (nextSystem) {
-                    score->systems().append(nextSystem);
-                }
-            }
-        } else {
-            nextSystem = score->collectSystem(*this);
-            if (nextSystem) {
-                collected = true;
-            }
-        }
-        prevSystem = curSystem;
-        Q_ASSERT(curSystem != nextSystem);
-        curSystem  = nextSystem;
-
-        bool breakPage = !curSystem || (breakPages && prevSystem->pageBreak());
-
-        if (!breakPage) {
-            qreal dist = prevSystem->minDistance(curSystem) + curSystem->height();
-            Box* vbox = curSystem->vbox();
-            if (vbox) {
-                dist += vbox->bottomGap();
-            } else if (!prevSystem->hasFixedDownDistance()) {
-                qreal margin = qMax(curSystem->minBottom(), curSystem->spacerDistance(false));
-                dist += qMax(margin, slb);
-            }
-            breakPage = (y + dist) >= ey && breakPages;
-        }
-        if (breakPage) {
-            qreal dist = qMax(prevSystem->minBottom(), prevSystem->spacerDistance(false));
-            dist = qMax(dist, slb);
-            layoutPage(page, ey - (y + dist));
-            // if we collected a system we cannot fit onto this page,
-            // we need to collect next page in order to correctly set system positions
-            if (collected) {
-                pageOldMeasure = nullptr;
-            }
-            break;
-        }
-    }
-
-    Fraction stick = Fraction(-1, 1);
-    for (System* s : page->systems()) {
-        Score* currentScore = s->score();
-        for (MeasureBase* mb : s->measures()) {
-            if (!mb->isMeasure()) {
-                continue;
-            }
-            Measure* m = toMeasure(mb);
-            if (stick == Fraction(-1, 1)) {
-                stick = m->tick();
-            }
-
-            for (int track = 0; track < currentScore->ntracks(); ++track) {
-                for (Segment* segment = m->first(); segment; segment = segment->next()) {
-                    Element* e = segment->element(track);
-                    if (!e) {
-                        continue;
-                    }
-                    if (e->isChordRest()) {
-                        if (!currentScore->staff(track2staff(track))->show()) {
-                            continue;
-                        }
-                        ChordRest* cr = toChordRest(e);
-                        if (notTopBeam(cr)) {                           // layout cross staff beams
-                            cr->beam()->layout();
-                        }
-                        if (notTopTuplet(cr)) {
-                            // fix layout of tuplets
-                            DurationElement* de = cr;
-                            while (de->tuplet() && de->tuplet()->elements().front() == de) {
-                                Tuplet* t = de->tuplet();
-                                t->layout();
-                                de = t;
-                            }
-                        }
-
-                        if (cr->isChord()) {
-                            Chord* c = toChord(cr);
-                            for (Chord* cc : c->graceNotes()) {
-                                if (cc->beam() && cc->beam()->elements().front() == cc) {
-                                    cc->beam()->layout();
-                                }
-                                cc->layoutSpanners();
-                                for (Element* element : cc->el()) {
-                                    if (element->isSlur()) {
-                                        element->layout();
-                                    }
-                                }
-                            }
-                            c->layoutArpeggio2();
-                            c->layoutSpanners();
-                            if (c->tremolo()) {
-                                Tremolo* t = c->tremolo();
-                                Chord* c1 = t->chord1();
-                                Chord* c2 = t->chord2();
-                                if (t->twoNotes() && c1 && c2 && (c1->staffMove() || c2->staffMove())) {
-                                    t->layout();
-                                }
-                            }
-                        }
-                    } else if (e->isBarLine()) {
-                        toBarLine(e)->layout2();
-                    }
-                }
-            }
-            m->layout2();
-        }
-    }
-
-    if (score->systemMode()) {
-        System* s = page->systems().last();
-        qreal height = s ? s->pos().y() + s->height() + s->minBottom() : page->tm();
-        page->bbox().setRect(0.0, 0.0, score->loWidth(), height + page->bm());
-    }
-
-    page->rebuildBspTree();
-}
-
-//---------------------------------------------------------
 //   doLayout
 //    do a complete (re-) layout
 //---------------------------------------------------------
@@ -5165,255 +1966,9 @@ void Score::doLayout()
     doLayoutRange(Fraction(0, 1), Fraction(-1, 1));
 }
 
-//---------------------------------------------------------
-//   CmdStateLocker
-//---------------------------------------------------------
-
-class CmdStateLocker
-{
-    Score* score;
-public:
-    CmdStateLocker(Score* s)
-        : score(s) { score->cmdState().lock(); }
-    ~CmdStateLocker() { score->cmdState().unlock(); }
-};
-
-//---------------------------------------------------------
-//   doLayoutRange
-//---------------------------------------------------------
-
 void Score::doLayoutRange(const Fraction& st, const Fraction& et)
 {
-    CmdStateLocker cmdStateLocker(this);
-    LayoutContext lc(this);
-
-    Fraction stick(st);
-    Fraction etick(et);
-    Q_ASSERT(!(stick == Fraction(-1, 1) && etick == Fraction(-1, 1)));
-
-    if (!last() || (lineMode() && !firstMeasure())) {
-        qDebug("empty score");
-        qDeleteAll(_systems);
-        _systems.clear();
-        qDeleteAll(pages());
-        pages().clear();
-        lc.getNextPage();
-        return;
-    }
-//      if (!_systems.isEmpty())
-//            return;
-    bool layoutAll = stick <= Fraction(0, 1) && (etick < Fraction(0, 1) || etick >= masterScore()->last()->endTick());
-    if (stick < Fraction(0, 1)) {
-        stick = Fraction(0, 1);
-    }
-    if (etick < Fraction(0, 1)) {
-        etick = last()->endTick();
-    }
-
-    lc.endTick = etick;
-    _scoreFont = ScoreFont::fontByName(style().value(Sid::MusicalSymbolFont).toString());
-    _noteHeadWidth = _scoreFont->width(SymId::noteheadBlack, spatium() / SPATIUM20);
-
-    if (cmdState().layoutFlags & LayoutFlag::REBUILD_MIDI_MAPPING) {
-        if (isMaster()) {
-            masterScore()->rebuildMidiMapping();
-        }
-    }
-    if (cmdState().layoutFlags & LayoutFlag::FIX_PITCH_VELO) {
-        updateVelo();
-    }
-#if 0 // TODO: needed? It was introduced in ab9774ec4098512068b8ef708167d9aa6e702c50
-    if (cmdState().layoutFlags & LayoutFlag::PLAY_EVENTS) {
-        createPlayEvents();
-    }
-#endif
-
-    //---------------------------------------------------
-    //    initialize layout context lc
-    //---------------------------------------------------
-
-    MeasureBase* m = tick2measure(stick);
-    if (m == 0) {
-        m = first();
-    }
-    // start layout one measure earlier to handle clefs and cautionary elements
-    if (m->prevMeasureMM()) {
-        m = m->prevMeasureMM();
-    } else if (m->prev()) {
-        m = m->prev();
-    }
-    while (!m->isMeasure() && m->prev()) {
-        m = m->prev();
-    }
-
-    // if the first measure of the score is part of a multi measure rest
-    // m->system() will return a nullptr. We need to find the multi measure
-    // rest which replaces the measure range
-
-    if (!m->system() && m->isMeasure() && toMeasure(m)->hasMMRest()) {
-        qDebug("  don’t start with mmrest");
-        m = toMeasure(m)->mmRest();
-    }
-
-//      qDebug("start <%s> tick %d, system %p", m->name(), m->tick(), m->system());
-
-    if (lineMode()) {
-        lc.prevMeasure = 0;
-        lc.nextMeasure = m;         //_showVBox ? first() : firstMeasure();
-        lc.startTick   = m->tick();
-        layoutLinear(layoutAll, lc);
-        return;
-    }
-    if (!layoutAll && m->system()) {
-        System* system  = m->system();
-        int systemIndex = _systems.indexOf(system);
-        lc.page         = system->page();
-        lc.curPage      = pageIdx(lc.page);
-        if (lc.curPage == -1) {
-            lc.curPage = 0;
-        }
-        lc.curSystem   = system;
-        lc.systemList  = _systems.mid(systemIndex);
-
-        if (systemIndex == 0) {
-            lc.nextMeasure = _showVBox ? first() : firstMeasure();
-        } else {
-            System* prevSystem = _systems[systemIndex - 1];
-            lc.nextMeasure = prevSystem->measures().back()->next();
-        }
-
-        _systems.erase(_systems.begin() + systemIndex, _systems.end());
-        if (!lc.nextMeasure->prevMeasure()) {
-            lc.measureNo = 0;
-            lc.tick      = Fraction(0, 1);
-        } else {
-            const MeasureBase* mb = lc.nextMeasure->prev();
-            if (mb) {
-                mb = mb->findPotentialSectionBreak();
-            }
-            LayoutBreak* sectionBreak = mb->sectionBreakElement();
-            // TODO: also use mb in else clause here?
-            // probably not, only actual measures have meaningful numbers
-            if (sectionBreak && sectionBreak->startWithMeasureOne()) {
-                lc.measureNo = 0;
-            } else {
-                lc.measureNo = lc.nextMeasure->prevMeasure()->no()                             // will be adjusted later with respect
-                               + (lc.nextMeasure->prevMeasure()->irregular() ? 0 : 1);         // to the user-defined offset.
-            }
-            lc.tick      = lc.nextMeasure->tick();
-        }
-    } else {
-//  qDebug("layoutAll, systems %p %d", &_systems, int(_systems.size()));
-        //lc.measureNo   = 0;
-        //lc.tick        = 0;
-        // qDeleteAll(_systems);
-        // _systems.clear();
-        // lc.systemList  = _systems;
-        // _systems.clear();
-
-        for (System* s : qAsConst(_systems)) {
-            for (Bracket* b : s->brackets()) {
-                if (b->selected()) {
-                    _selection.remove(b);
-                    setSelectionChanged(true);
-                }
-            }
-//                  for (SpannerSegment* ss : s->spannerSegments())
-//                        ss->setParent(0);
-            s->setParent(nullptr);
-        }
-        for (MeasureBase* mb = first(); mb; mb = mb->next()) {
-            mb->setSystem(0);
-            if (mb->isMeasure() && toMeasure(mb)->mmRest()) {
-                toMeasure(mb)->mmRest()->setSystem(0);
-            }
-        }
-        qDeleteAll(_systems);
-        _systems.clear();
-
-        qDeleteAll(pages());
-        pages().clear();
-
-        lc.nextMeasure = _showVBox ? first() : firstMeasure();
-    }
-
-    lc.prevMeasure = 0;
-
-    getNextMeasure(lc);
-    lc.curSystem = collectSystem(lc);
-
-    lc.layout();
-}
-
-//---------------------------------------------------------
-//   layout
-//---------------------------------------------------------
-
-void LayoutContext::layout()
-{
-    MeasureBase* lmb;
-    do {
-        getNextPage();
-        collectPage();
-
-        if (page && !page->systems().isEmpty()) {
-            lmb = page->systems().back()->measures().back();
-        } else {
-            lmb = nullptr;
-        }
-
-        // we can stop collecting pages when:
-        // 1) we reach the end of score (curSystem is nullptr)
-        // or
-        // 2) we have fully processed the range and reached a point of stability:
-        //    a) we have completed layout for the range (rangeDone is true)
-        //    b) we haven't collected a system that will need to go on the next page
-        //    c) this page ends with the same measure as the previous layout
-        //    pageOldMeasure will be last measure from previous layout if range was completed on or before this page
-        //    it will be nullptr if this page was never laid out or if we collected a system for next page
-    } while (curSystem && !(rangeDone && lmb == pageOldMeasure));
-    // && page->system(0)->measures().back()->tick() > endTick // FIXME: perhaps the first measure was meant? Or last system?
-
-    if (!curSystem) {
-        // The end of the score. The remaining systems are not needed...
-        qDeleteAll(systemList);
-        systemList.clear();
-        // ...and the remaining pages too
-        while (score->npages() > curPage) {
-            delete score->pages().takeLast();
-        }
-    } else {
-        Page* p = curSystem->page();
-        if (p && (p != page)) {
-            p->rebuildBspTree();
-        }
-    }
-    score->systems().append(systemList);       // TODO
-}
-
-//---------------------------------------------------------
-//   LayoutContext::LayoutContext
-//---------------------------------------------------------
-
-LayoutContext::LayoutContext(Score* s)
-    : score(s)
-{
-    firstSystemIndent = score && score->styleB(Sid::enableIndentationOnFirstSystem);
-}
-
-//---------------------------------------------------------
-//   LayoutContext::~LayoutContext
-//---------------------------------------------------------
-
-LayoutContext::~LayoutContext()
-{
-    for (Spanner* s : processedSpanners) {
-        s->layoutSystemsDone();
-    }
-
-    for (MuseScoreView* v : score->getViewer()) {
-        v->layoutChanged();
-    }
+    m_layout.doLayoutRange(st, et);
 }
 
 //---------------------------------------------------------
