+++ src/engraving/libmscore/layout.h
@@ -87,50 +87,6 @@ public:
 };
 
 //---------------------------------------------------------
-//   LayoutContext
-//    temp values used during layout
-//---------------------------------------------------------
-
-struct LayoutContext {
-    Score* score             { 0 };
-    bool startWithLongNames  { true };
-    bool firstSystem         { true };
-    bool firstSystemIndent   { true };
-    Page* page               { 0 };
-    int curPage              { 0 };        // index in Score->page()s
-    Fraction tick            { 0, 1 };
-
-    QList<System*> systemList;            // reusable systems
-    std::set<Spanner*> processedSpanners;
-
-    System* prevSystem       { 0 };       // used during page layout
-    System* curSystem        { 0 };
-
-    MeasureBase* systemOldMeasure { 0 };
-    MeasureBase* pageOldMeasure   { 0 };
-    bool rangeDone           { false };
-
-    MeasureBase* prevMeasure { 0 };
-    MeasureBase* curMeasure  { 0 };
-    MeasureBase* nextMeasure { 0 };
-    int measureNo            { 0 };
-    Fraction startTick;
-    Fraction endTick;
-
-    LayoutContext(Score* s);
-    LayoutContext(const LayoutContext&) = delete;
-    LayoutContext& operator=(const LayoutContext&) = delete;
-    ~LayoutContext();
-
-    void layoutLinear();
-
-    void layout();
-    int adjustMeasureNo(MeasureBase*);
-    void getNextPage();
-    void collectPage();
-};
-
-//---------------------------------------------------------
 //   VerticalAlignRange
 //---------------------------------------------------------
 
@@ -138,9 +94,7 @@ enum class VerticalAlignRange {
     SEGMENT, MEASURE, SYSTEM
 };
 
-extern bool isTopBeam(ChordRest* cr);
 extern bool notTopBeam(ChordRest* cr);
-extern bool isTopTuplet(ChordRest* cr);
 extern bool notTopTuplet(ChordRest* cr);
 }     // namespace Ms
 #endif
