+++ src/engraving/libmscore/layout/layoutmeasure.h
@@ -19,36 +19,22 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-
-#include "score.h"
-#include "page.h"
-#include "system.h"
-#include "tremolo.h"
-#include "measure.h"
-#include "layout.h"
-#include "bracket.h"
-#include "spanner.h"
-#include "barline.h"
-#include "tie.h"
-#include "chord.h"
-#include "staff.h"
-#include "box.h"
-#include "spacer.h"
-#include "systemdivider.h"
-#include "tuplet.h"
-#include "dynamic.h"
-#include "stafflines.h"
-#include "tempotext.h"
-#include "hairpin.h"
-#include "part.h"
-#include "keysig.h"
-#include "sig.h"
-#include "breath.h"
-#include "tempo.h"
-#include "fermata.h"
-#include "lyrics.h"
-
-using namespace mu;
+#ifndef MU_ENGRAVING_LAYOUTMEASURE_H
+#define MU_ENGRAVING_LAYOUTMEASURE_H
 
 namespace Ms {
-} // namespace Ms
+class Score;
+}
+
+namespace mu::engraving {
+class LayoutContext;
+class LayoutMeasure
+{
+public:
+    LayoutMeasure() = default;
+
+    static void getNextMeasure(Ms::Score* score, LayoutContext& lc);
+};
+}
+
+#endif // MU_ENGRAVING_LAYOUTMEASURE_H
