+++ src/engraving/libmscore/libmscore.cmake
@@ -346,4 +346,15 @@ set(LIBMSCORE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/vibrato.h
     ${CMAKE_CURRENT_LIST_DIR}/volta.cpp
     ${CMAKE_CURRENT_LIST_DIR}/volta.h
+
+    ${CMAKE_CURRENT_LIST_DIR}/layout/layout.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/layout/layout.h
+    ${CMAKE_CURRENT_LIST_DIR}/layout/layoutcontext.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/layout/layoutcontext.h
+    ${CMAKE_CURRENT_LIST_DIR}/layout/layoutlyrics.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/layout/layoutlyrics.h
+    ${CMAKE_CURRENT_LIST_DIR}/layout/layoutmeasure.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/layout/layoutmeasure.h
+    ${CMAKE_CURRENT_LIST_DIR}/layout/layoutbeams.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/layout/layoutbeams.h
 )
