+++ src/engraving/libmscore/tremolo.cpp
@@ -29,6 +29,8 @@
 #include "style/style.h"
 #include "io/xml.h"
 
+#include "layout/layout.h"
+
 #include "score.h"
 #include "staff.h"
 #include "chord.h"
@@ -309,8 +311,6 @@ void Tremolo::layoutOneNoteTremolo(qreal x, qreal y, qreal spatium)
     setPos(x, y);
 }
 
-extern std::pair<qreal, qreal> extendedStemLenWithTwoNoteTremolo(Tremolo*, qreal, qreal);
-
 //---------------------------------------------------------
 //   layoutTwoNotesTremolo
 //---------------------------------------------------------
@@ -359,7 +359,7 @@ void Tremolo::layoutTwoNotesTremolo(qreal x, qreal y, qreal h, qreal spatium)
     } else {
         firstChordStaffY = _chord1->pagePos().y() - _chord1->y();      // y coordinate of the staff of the first chord
         const std::pair<qreal, qreal> extendedLen
-            = extendedStemLenWithTwoNoteTremolo(this, _chord1->defaultStemLength(), _chord2->defaultStemLength());
+            = mu::engraving::Layout::extendedStemLenWithTwoNoteTremolo(this, _chord1->defaultStemLength(), _chord2->defaultStemLength());
         y1 = _chord1->stemPos().y() - firstChordStaffY + extendedLen.first;
         y2 = _chord2->stemPos().y() - firstChordStaffY + extendedLen.second;
     }
